==> prep/03_prep_spp_habitats/01_download_AOH_terrestrial.Rmd <==
---
title: "Download terrestrial Area of Habitat (AOH) data"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary

Here we download AOH maps from Eyres et al and Lumbierres et al (reference below).

## Data sources 

* Eyres et al. 2023: preprint: https://www.researchgate.net/publication/376637364_LIFE_A_metric_for_quantitively_mapping_the_impact_of_land-cover_change_on_global_extinctions Files shared through email communication with Alison Eyres (ae491@cam.ac.uk) and Michael Winston Dales (mwd24@cam.ac.uk) at Cambridge University: https://www.cl.cam.ac.uk/research/eeg/4c/data/aoh/
* Lumbierres et al. 2022: https://www.nature.com/articles/s41597-022-01838-w


## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(terra)
library(strex)
library(janitor)
library(rredlist)
library(sf)
library(mapview)

source(here("src/directories.R"))
source(here("src/spatial.R"))
source(here("src/fxns.R"))

select <- dplyr::select
setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # set working directory to where this script is located
this_dir <- getwd()

```

## Download the Eyres data

NOTE: This link and data were provided by Alison Eyres at Cambridge (ae491@cam.ac.uk)

run these bash in the terminal and you will be able to recursively download all species AOH files from https://www.cl.cam.ac.uk/research/eeg/4c/data/aoh/ to your specific folder (mine is "AOH_eyres")

IUCN ids are the numbers which are attached

This takes a really long time... run it overnight.

```{bash eval = false}

cd ../../raw_data/AOH_eyres
wget -r -nH --cut-dirs=5 --no-parent --reject="index.html*" -N https://www.cl.cam.ac.uk/research/eeg/4c/data/aoh/

```

Get IUCN ids and species names for the Eyres species (need to download them all first)

```{r}

out <- rl_sp(all=TRUE, key = api_key)
vapply(out,"[[",1,"count") 
all_df<-do.call(rbind,lapply(out,"[[","result"))

write_rds(all_df, here("prep/03_prep_spp_habitats/data/iucn/all_iucn_spp.rds"))


spp_ids <- data.frame(fn = grep(list.files(file.path("/mnt/rdsi/raw_data/AOH_eyres/"), recursive = TRUE), pattern = "reprojected", invert = TRUE, value = TRUE)) %>% # delete amphibians and add recursive for all species names
  mutate(iucn_id = as.numeric(str_after_last(fn, "-") %>%
  str_before_first(., ".tif")),
  spp_type = str_before_first(fn, "\\/")) %>%
  distinct(spp_type, iucn_id)

all_df <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/all_iucn_spp.rds"))

eyres_ids <- spp_ids %>% 
  left_join(all_df, by = c("iucn_id" = "taxonid")) # ok cool 

spp_groups <- eyres_ids %>%
  group_by(spp_type) %>%
  summarise(n_group = n_distinct(iucn_id)) %>%
  ungroup()

test_nas <- eyres_ids %>%
  filter(is.na(scientific_name)) %>% ## missing names for 235 species
  dplyr::select(spp_type, iucn_id)

write.csv(test_nas, here("prep/03_prep_spp_habitats/data/iucn/missing_ids.csv"), row.names = FALSE)

# lets see if we can find some of these names using iucn redlist api 

loop_list <- c()

for(loop_id in unique(test_nas$iucn_id)){
  
  search_id <- rl_search(id = loop_id, key = api_key)
  
  loop_list <- c(loop_list, search_id)
} # ok that didn't work.. maybe these just don't exist? 

write_rds(eyres_ids, here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds"))

```


Data check after reprojection in `02_map_terrestrial_to_moll.R`

```{r}
new <- read.csv("/mnt/rdsi/raw_data/AOH_eyres/reprojected_mol_csv/Seasonality.RESIDENT-88149202.csv")

new1 <- new %>% 
    left_join(moll_template_xy) %>%
  dplyr::select(x, y, presence) %>%
  rast(., type = "xyz", crs = moll_template)
plot(new1)

      
old <- raster("/mnt/rdsi/raw_data/AOH_eyres/mammals/Seasonality.RESIDENT-88149202.tif") 
# %>% 
# projectRaster(., moll_template, method = "ngb")

# there are 0s in the files... is using nearest neighbor screwing with some of the smaller files? I'm not really sure... the big files seem ok. And taking out all of the 0s take forever. 

# %>%
# classify(., cbind(-Inf, 0, NA), right=FALSE) # clamp, ifel, and classify in terra and raster all take a really long time...
# clamp(., lower=0.000001, useValues=FALSE)
# old[old == 0] <- NA

plot(old)  

# ok so big files seem to be ok... lets look at small files


new <- read.csv("/mnt/rdsi/raw_data/AOH_eyres/reprojected_mol_csv/Seasonality.RESIDENT-77347251.csv")

new1 <- new %>% 
    left_join(moll_template_xy) %>%
  dplyr::select(x, y, presence) %>%
  st_as_sf(., coords = c("x", "y"), crs = crs(moll_template))

mapview(new1)

      
old <- rast("/mnt/rdsi/raw_data/AOH_eyres/amphibians/Seasonality.RESIDENT-77347251.tif")     
# old      
plot(old)  

old1 <- old %>%
        aggregate(fact = 10, fun = "mean", na.rm = TRUE) %>%  
        project(., moll_template, method = "near") %>%
  as.data.frame(., xy = TRUE) %>%
  rename(presence = 3) %>%
  filter(presence != 0) %>%
  mutate(presence = 1 ) %>%
  left_join(moll_template_xy) %>%
  dplyr::select(x, y, presence) %>%
  rast(., type = "xyz")

plot(old1)

# Let's rerun everything under 300 bytes with the aggregate method just to be sure. I don't think it will really make a difference. A few cells here and there. It is definitely working... just not as exact as I would like. 

```


ARCHIVE: 


## Download the Lumbierres data 
 - NOTE: you don't need to do this unless you plan to use the Lumbierres data. The Eyres data is more up-to-date and comprehensive.

Check to make sure we downloaded all of the AOH species maps from https://datadryad.org/stash/dataset/doi:10.5061/dryad.02v6wwq48


You can download by ssh into the rdsi storage and then running wget url_to_download

ssh ubuntu@131.217.178.173
cd ../../mnt/rdsi/raw_data/AOH_lumbierres
wget https://datadryad.org/stash/downloads/file_stream/1931890

unzip 1931890.zip

Above example is how to download mammals_rodentia.zip file. Link taken from https://datadryad.org/stash/dataset/doi:10.5061/dryad.02v6wwq48.


```{r}

## read in species lists from raw data dir 

mammals_list <-
  read.csv(file.path(rdsi_raw_data_dir, "AOH_lumbierres/Mammals_list_AOH.csv")) %>%
  clean_names() %>%
  dplyr::select(binomial, aoh_name, order) %>%
  mutate(type = "mammal")

birds_list <- read.csv(file.path(rdsi_raw_data_dir, "AOH_lumbierres/Birds_list_AOH.csv")) %>%
  clean_names() %>%
  dplyr::select(binomial, aoh_name, order) %>%
  mutate(type = "bird")


total_spp_list <- rbind(mammals_list, birds_list)

## save species list 
write_rds(total_spp_list, file.path(this_dir, "int/terrestrial_species_list.rds"))


## now grab the names of the species from the raw data directory

files <- list.files(file.path(rdsi_raw_data_dir, "AOH_lumbierres"), pattern = ".tif")

spp_list_files <- str_before_last(files, ".tif") 

remove_suffix <- function(string) {
  modified_string <- sub("_[BNR]$", "", string)
  return(modified_string)
}

spp_list_files_fin <- remove_suffix(spp_list_files) %>%
  unique()


## now filter these spp out of the spp df 

missing_spp <- total_spp_list %>%
  filter(!(aoh_name %in% c(spp_list_files_fin)))

## ok I'm missing 1984 spp. 

sort(unique(missing_spp$order)) 

# [1] "CHIROPTERA"    "PASSERIFORMES"    


## I will redownload the zip files missing from here: https://datadryad.org/stash/dataset/doi:10.5061/dryad.02v6wwq48 and see if that helps the missing species...

unzip(file.path(rdsi_raw_data_dir, "AOH_lumbierres/1931870.zip"), exdir = here("test/"))


test <- rast(file.path(rdsi_raw_data_dir, "AOH_lumbierres/Aselliscus_tricuspidatus.tif"))

test_size <- cellSize(test, unit = "km")
plot(test_size)


```
