==> prep/03_prep_spp_habitats/04_terrestrial_overlap_summary.Rmd <==
---
title: "Summarizing global terrestrial AOH impact per species"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary

In this script we prep data frames of total global amount of impacted habitat per terrestrial species.


## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(terra)
library(parallel)
library(strex)
library(janitor)
library(sf)
library(tools)
library(glue)

source(here("src/directories.R"))

source(here("src/spatial.R"))

aoh_dir <- file.path(rdsi_raw_data_dir, "AOH_eyres/reprojected_mol_csv_fin/")

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")
```

## Methods 

First, save habitat ares for each species. Additionally, find the habitat areas for any species that are missing (3617 of them, because they have no exposure, and thus, no impact)

```{r}

impact_df <-  purrr::map_df(list.files(file.path(biodiv_dir, "int/aoh_impacts_terrestrial/"), full.names = TRUE), readRDS) %>%
  rename(sciname = species_full) %>%
  mutate(sensitivity_scenario = ifelse(is.na(sensitivity_scenario), "original", sensitivity_scenario))

length(unique(impact_df$sciname)) # 27911 - missing some bc no exposure

spp_habs_df <- impact_df %>%
  distinct(sciname, total_hab_area = hab_area)

  
## get sensitivity info dataframe
spp_types <- readRDS(file.path(here("prep/03_prep_spp_habitats"), "int/terrestrial_spp_habitat_suitability.rds")) %>%
  dplyr::distinct(sciname = species, spp_type)

# now lets clean up the data a bit 

all_spp_list <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(sciname = scientific_name, order = order_name, iucn_rl_cat = category) %>%
  filter(!is.na(sciname)) %>%
  distinct() ## ok this is the correct number of species

## what species are missing? 

missing_spp <- setdiff(all_spp_list$sciname, spp_habs_df$sciname) # ~3685 species missing

## get their species IDs so we can read in the maps and save total habitat area
all_iucn_spp <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/all_iucn_spp.rds"))

missing_iucn_spp <- all_iucn_spp %>%
  filter(scientific_name %in% missing_spp) %>%
  distinct(taxonid) %>%
  pull() # grab the taxon id so we can loop through the csvs


missing_spp_habs <- data.frame(taxonid = NA, total_hab_area = NA)

for(spp_id in missing_iucn_spp){
  
  # spp_id = 58058 # ok so, the missing ones all seem to have realllllly small habitats, and that's why they don't show. We will just save those as having 100km2 of habitat as this is the minimum amount a species could have with our resolution... And assume no habitat area impact.
  
  spp_rast <- read.csv(glue(file.path(aoh_dir, "{spp_id}.csv")))
  
  if(nrow(spp_rast) == 0){
    
    spp_hab <- data.frame(taxonid = spp_id, total_hab_area = 100)
    
  }else{ # i don't think there are any "else" instances but just in case
  
  total_hab <- sum(spp_rast$presence)*100
  
  spp_hab <- data.frame(taxonid = spp_id, total_hab_area = total_hab)
    
  }
  
  
  missing_spp_habs <- rbind(missing_spp_habs, spp_hab)
  
}

# join with iucn data to get sciname
missing_spp_fin <- missing_spp_habs %>%
  left_join(all_iucn_spp) %>%
  filter(!is.na(total_hab_area)) %>%
  dplyr::select(sciname = scientific_name, total_hab_area)


spp_habs_df_fin <- spp_habs_df %>%
  rbind(., missing_spp_fin)

write.csv(spp_habs_df_fin, here("prep/03_prep_spp_habitats/int/terrestrial_habitat_areas.csv"), row.names = FALSE)

test <- data.frame(dups = duplicated(spp_habs_df_fin$sciname)) %>%
  filter(dups == TRUE) # good, 0 duplicates

```

Bind together all habitat impact data from the overlap script

```{r}

## need to do some sort of expand.grid with missing species and all fcr/diet scenarios

diets <- unique(impact_df$diet)
fcr_types <- unique(impact_df$fcr_type)
allocations <- unique(impact_df$allocation)
crop_ingredients <- unique(impact_df$crop_ingredient)
sensitivity_scenarios <- unique(impact_df$sensitivity_scenario)

expand_df <- expand_grid(missing_spp_fin, diet = diets, fcr_type = fcr_types, allocation = allocations, crop_ingredient = crop_ingredients, sensitivity_scenario = sensitivity_scenarios) %>%
  mutate(impact_total = 0) %>%
  dplyr::select(-total_hab_area)

impact_df_tidy <- impact_df %>%
    group_by(sciname, diet, fcr_type, allocation, crop_ingredient, sensitivity_scenario) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  rbind(., expand_df) %>%
  mutate(aoh_filepath = paste0(aoh_dir, sciname), 
         feed_filepath = sprintf(here("prep/02_feed/output/resampled/%s/%s/%s_%s_A.tif"), diet, fcr_type, crop_ingredient, allocation)) %>%
  rename("ingredient" = "crop_ingredient") %>%
  mutate(sciname = str_replace_all(sciname, "_", " ")) %>%
  left_join(all_spp_list, by = c("sciname")) %>%
  left_join(spp_habs_df_fin, by = "sciname") %>%
  mutate(impact_total = ifelse(is.na(impact_total), 0, impact_total))


species_list <- c(unique(impact_df_tidy$sciname))

## save lists of assessed vs impacted species 
writeLines(species_list, here("prep/03_prep_spp_habitats/int/terr_assessed_species_list.txt"))

writeLines(impact_df_tidy %>% filter(impact_total > 0) %>% pull(sciname) %>% unique(), here("prep/03_prep_spp_habitats/int/terr_impacted_species_list.txt"))

hab_areas <- impact_df_tidy %>% ungroup() %>% distinct(sciname, total_hab_area) %>%
  group_by(sciname) %>%
  filter(total_hab_area == max(total_hab_area)) %>%
  ungroup()

test <- impact_df_tidy %>%
  ungroup() %>%
  group_by(sciname, diet, fcr_type, allocation) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  left_join(spp_habs_df) %>% 
  mutate(prop_impacted = impact_total/total_hab_area)

## highest total proportional impact is Liolaemus gravenhorstii, commonly known as Gravenhorst's tree iguana. ~1.1% of its habitat is impacted under plant-dominant, regular, economic. 
```


Bind species' sensitivities to the full aoh impacted dataframe. 

```{r}

## get sensitivity info dataframe
ag_suitability <- readRDS(file.path(here("prep/03_prep_spp_habitats"), "int/terrestrial_spp_habitat_suitability.rds")) %>%
  mutate(cropland_sensitivity = 1 - cropland_suitability) %>%
  mutate(cropland_sensitivity_plus = case_when(cropland_sensitivity == 0 ~ 0.1,
                                               cropland_sensitivity == 0.5 ~ 0.6, 
                                               cropland_sensitivity == 1 ~ 1,
                                               TRUE ~ cropland_sensitivity)) %>%
  mutate(cropland_sensitivity_minus = case_when(cropland_sensitivity == 0 ~ 0,
                                                cropland_sensitivity == 0.5 ~ 0.4, 
                                                cropland_sensitivity == 1 ~ 0.9,
                                               TRUE ~ cropland_sensitivity)) %>%
  pivot_longer(cols = c(cropland_sensitivity, cropland_sensitivity_plus, cropland_sensitivity_minus), names_to = "sensitivity_scenario", values_to = "vulnerability") %>%
  mutate(sensitivity_scenario = case_when(
    sensitivity_scenario == "cropland_sensitivity" ~ "original",
    sensitivity_scenario == "cropland_sensitivity_plus" ~ "plus-10",
    sensitivity_scenario == "cropland_sensitivity_minus" ~ "minus-10",
  )) %>%
  dplyr::select(-spp_type, -cropland_suitability)

test <- ag_suitability %>%
  filter(species %in% missing_spp) 

test <- ag_suitability %>%
  filter(vulnerability > 0) %>%
  group_by(sensitivity_scenario) %>% 
  summarise(count = n()) %>% 
  ungroup()

aoh_df <- impact_df_tidy %>%
  left_join(., ag_suitability, by = c("sciname" = "species", "sensitivity_scenario")) %>%
  left_join(spp_types)

saveRDS(aoh_df, here("prep/03_prep_spp_habitats/output/aoh_terrestrial_impact.rds")) # save as rds to save space

```

Data checking 

```{r}
length(unique(all_spp_list$sciname)) # 31596
length(unique(aoh_df$sciname)) # 31596
# perfect


## paper reports: 

length(unique(ag_suitability$species)) # 31596
ag_suitability %>% filter(vulnerability != 0, sensitivity_scenario == "original") %>% pull(species) %>% unique() %>% length() # 26413

## filter for our missing spp and see their suitability

test_ag <- ag_suitability %>% 
  filter(species %in% c(missing_spp)) # ok... majority of them should theoretically be impacted... does this just mean there was literally no overlap at all? Yes


all_spp_list <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(sciname = scientific_name, order = order_name, iucn_rl_cat = category, spp_type) %>%
  group_by(spp_type) %>%
  summarise(n_distinct(sciname))

all_spp_list

# # A tibble: 4 × 2
#   spp_type   `n_distinct(sciname)`
#   <chr>                      <int>
# 1 amphibians                  6793
# 2 birds                      10025
# 3 mammals                     5502
# 4 reptiles                    9280

```

Check soybean impacts:

```{r}

soybean_imps <- aoh_df %>%
  filter(fcr_type == "regular",
         str_detect(ingredient, "Soybean"),
         allocation == "economic") %>%
  group_by(diet, fcr_type, allocation, ingredient, sensitivity_scenario) %>%
  summarise(total_impact = sum(impact_total, na.rm = TRUE)) %>%
  ungroup()

```
