==> prep/03_prep_spp_habitats/05a_prep_aquamaps_depths.Rmd <==
---
title: "Prep aquamaps depths"
name: "Gage Clawson"
date: "September 25, 2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary 

In this script we find depth information for all AquaMaps species. 
    
 - Categorize each Aquamaps species into depth zones of reef, pelagic, bentho-pelagic, or benthic. This code was adapted from Casey O'Hara (O'Hara et al. 2023, in prep), and pulls from Fishbase, Sealifebase, and Butt et al. 2022 to obtain the appropriate depth ranges; https://github.com/mapping-marine-spp-vuln/spp_vuln_mapping/tree/master
 - This is necessary for the next script, in which we clip downscaled and re-projected AquaMaps species suitable habitat maps (>0.6 probability) (from 0.5 to ~10km resolution) to a bathymetric layer, so as to constrain neritic and shallow-water species to areas no deeper than 200 m (adapted from O'Hara et al. 2021). It is also necessary for the script in which we overlap and calculate habitat impact, as some species may not be exposed to certain gear types of catch based on their depth preferences. Basically, we are trying to introduce a 3D component while still using 2D AOH raster maps. Save each as a csv with cell id and presence/absence column


## References

* Froese, R. and D. Pauly, Editors. 2000. FishBase 2000: concepts, design and data sources. ICLARM, Los Baños, Laguna, Philippines. 344 p.
* Houde, E.D. and C.E. Zastrow. 1993. Ecosystem- and taxon-specific dynamic energetics properties of fish larvae assemblages. Bull. Mar. Sci. 53(2):290-335.
* Sa-a, P., M.L. Palomares and D. Pauly. 2000. The FOOD ITEMS table, p. 182-188. In R. Froese and D. Pauly (eds.) FishBase 2000: concepts, design and data sources. ICLARM, Los Baños, Laguna, Philippines. 344 p.
* Welcomme, R.L. 1988. International introductions of inland aquatic species. FAO Fish. Tech. Pap. 294, 318 p.
* Butt, N. et al. 2022.
* Kaschner, K., Kesner-Reyes, K., Garilao, C., Segschneider, J., Rius-Barile, J. Rees, T., & Froese, R. (2019, October). AquaMaps: Predicted range maps for aquatic species. Retrieved from https://www.aquamaps.org.
* O'Hara et al. 2023 in prep 
* Casey C. O’Hara et al., At-risk marine biodiversity faces extensive, expanding, and intensifying human impacts.Science372,84-87(2021).DOI:10.1126/science.abe6731

FishBase is a scientific database, and this has the implication - among others - that its use and the use of its contents are free as long as due credit is given.

This may be done at different levels, for which we suggest different forms of citations:

* when referring to FishBase concepts and design, cite its architects (Froese and Pauly 2000);
* when referring to a set of values extracted from a FishBase table, cite the author(s) of the original data, e.g., "Houde and Zastrow (1993)", or "Welcomme (1988)". To help us track the use of FishBase in the literature, we would appreciate your also citing Froese and Pauly (2000) in an appropriate part of your text, as the source of the information;
* when discussing the features of a FishBase table, cite the section documenting that table, e.g., "Sa-a et al. (2000)."


## Setup 

```{r setup, include=FALSE}

library(tidyverse)
library(tidyr)
library(here)
library(sf)
library(data.table)
library(dtplyr)
library(terra)
library(parallel)
library(strex)
library(janitor)
library(readxl)
library(rfishbase)

source(here("src/directories.R"))

source(here("src/spatial.R"))

aquamaps_dir <- file.path(rdsi_raw_data_dir, "aquamaps")


gall_peters <- "+proj=cea +lon_0=0 +x_0=0 +y_0=0 +lat_ts=45 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
#raster template 
base_rast <- rast(res=0.5)
ext(base_rast) <- c(-180, 180, -90, 90)

```

Read in Aquamaps data and filter for prob > 0.6

```{r}

aquamaps_csv <- read.csv(file.path(aquamaps_dir, "GTE10_HSPEC_NATIVE/hcaf_species_native_gte10.csv"))

# colnames(aquamaps_csv)
# length(unique(aquamaps_csv$SpeciesID)) # 23699... each have a distinct id. README says that ids are from catalogueoflife.org

aquamaps_spp <- read.csv(file.path(aquamaps_dir, "GTE10_SPECIESOCCURSUM/speciesoccursum_gte10.csv")) %>%
  distinct(SpeciesID, SpecCode, Genus, Species, FBname)


aquamaps_spp_df <- aquamaps_csv %>%
  filter(Probability > 0.6) %>%
   left_join(aquamaps_spp, by = "SpeciesID")

rm(aquamaps_csv)


```


### Find depth information for all aquamaps species

Grab depth information (continuous depth values) that AquaMaps team provided. Then we need to classify each species into "reef associated", "pelagic", "bentho-pelagic", and "benthic", for exposure calculations. These depth classifications are based on depth classifications in Butt et al. 2022 and O'Hara et al. 2023 (in review).


```{r}
## grab continuous depth values; provided directly by Aquamaps team
aquamaps_depth <- read.csv(file.path(aquamaps_dir, "ver10_2019_hspen.csv"), sep="\t", header=TRUE) %>%
  dplyr::select(SpeciesID, SpecCode = Speccode, DepthPrefMin, DepthPrefMax) %>%
  right_join(aquamaps_spp) %>%
  mutate(species = paste0(tolower(Genus), " ", Species))

## now we need to classify those depth values into "reef", "pelagic", "bentho-pelagic", and "benthic". We can start this by reading in depth information from the trait-based assessment (Butt et al., 2022). Code taken from Casey O'Hara spp_vuln_mapping project: https://github.com/mapping-marine-spp-vuln/spp_vuln_mapping/blob/master/1_setup/functional_traits/1d_collect_water_col_pos.Rmd#L38

vuln_info <- read.csv(here("prep/03_prep_spp_habitats/data/spp_vuln_butt/spp_traits_valid.csv")) %>% # dataset downloaded from KNB: https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1FX77VJ
  filter(trait %in% c('zone', 'terrestrial_and_marine_life_stages', 
                      'within_stage_dependent_habitats_condition', 
                      'across_stage_dependent_habitats_condition', 
                      'air_sea_interface'))


btm_terms <- c('bottom', 'slope', 'shelf', 'mud', 
               'benth', 'rubble', 'intertid', 'demersal') %>%
  paste0(collapse = '|') # terms associated with bottom/benthic

reef_terms <- c('reef') %>%
  paste0(collapse = '|') # reef terms 

mid_terms <- c('pelagic', 'mid', 'ocean') %>%
  paste0(collapse = '|') # midwater terms, e.g., pelagic

depth_vuln_downfill <- vuln_info %>%
  group_by(taxon, spp_gp) %>%
  summarize(trait_value = paste0(trait_value, collapse = ';'), .groups = 'drop') %>%
  fill(.direction = "down") # downfill any missing depth values (I don't think there are any missing?)

depth_vuln_score <- depth_vuln_downfill %>%
  rowwise() %>%
  mutate(pelagic = any(str_detect(trait_value, mid_terms)),
         benthic = any(str_detect(trait_value, btm_terms)),
         reef    = any(str_detect(trait_value, reef_terms)),
         .groups = 'drop') %>% # categorize into benthic, pelagic, or reef
  ungroup() %>%
  mutate(reef = ifelse(taxon == 'sponges', TRUE, reef)) %>%
  dplyr::select(taxon, species = spp_gp, benthic, pelagic, reef) %>%
  distinct()


```

Get depth categorizations from Fishbase and Sealifebase 

```{r}

# first look in Fishbase

spp_list <- aquamaps_spp %>%
  unite("spp_name", Genus:Species, sep = " ") %>%
  pull(spp_name)

species_info_fb <- species(spp_list) %>%
  dplyr::select(species = Species, SpecCode, DemersPelag) %>%
  filter(!is.na(DemersPelag)) %>%
    mutate(benthic = str_detect(DemersPelag, 'benth|sessile|demersal'),
         pelagic = str_detect(DemersPelag, 'pelagic'),
         reef    = str_detect(DemersPelag, 'reef')) %>%
  dplyr::select(-DemersPelag) %>%
  filter(benthic | pelagic | reef) %>%
  mutate(taxon = NA) %>%
  mutate(species = tolower(species))

# now look in Sealifebase

species_info_slb <- species(spp_list, server = "sealifebase") %>%
  dplyr::select(species = Species, SpecCode, DemersPelag) %>%
  filter(!is.na(DemersPelag)) %>%
    mutate(benthic = str_detect(DemersPelag, 'benth|sessile|demersal'),
         pelagic = str_detect(DemersPelag, 'pelagic'),
         reef    = str_detect(DemersPelag, 'reef')) %>%
  dplyr::select(-DemersPelag) %>%
  filter(benthic | pelagic | reef) %>%
  mutate(taxon = NA) %>%
  mutate(species = tolower(species))

species_info <- rbind(species_info_slb, species_info_fb)


# rbind fishbase and vuln datasets
fb_vuln_depth_df <- depth_vuln_score %>%
  filter(!species %in% species_info$species) %>%
  mutate(source = 'vuln_traits') %>%
  bind_rows(species_info %>%
              mutate(source = 'fishbase')) %>%
  gather(value, bool, benthic, pelagic, reef) %>%
  filter(bool == TRUE) %>%
  dplyr::select(species, value, source, taxon) %>%
  mutate(genus = str_extract(species, '[a-z]+(?= ?)')) %>%
  distinct() 
  
```

**Now we need to join the depth category information with the aquamaps depth information and gapfill any missing categories**

For processing, consider trait categories (in order of preference, but not an ordered category) of benthopelagic, pelagic, reef, benthic.  Few reef species are pelagic; this order will assign those to benthopelagic (along with all others that are both benthic and pelagic).  

Remaining pelagic species will not be benthic or reef-associated.  

Some reef spp are also benthic, but reef association seems more likely to be important as a trait, so these will be assigned as reef. 

Because many species from FishBase/Butt vulnerability paper appear to have non-matched names from aquamaps (e.g., unaccepted synonyms), while many aquamaps species in the same genus are unscored, use genus-level (then family, order, class, phylum, and kingdom, in that order) gapfill to assign traits from unmatched species to aquamaps-matched species. Identify genera with more than one scored spp; select those traits that appear in the majority of the genus (e.g., use the mode).
 - Gapfilling is only necessary for ~3% of the species. 

```{r}

aquamaps_info <- read.csv(file.path(aquamaps_dir, "GTE10_SPECIESOCCURSUM/speciesoccursum_gte10.csv"))

aquamaps_depth_cats <- aquamaps_depth %>% 
  left_join(fb_vuln_depth_df) %>%
  mutate(genus = tolower(Genus)) %>%
  dplyr::select(-Genus) %>%
  left_join(aquamaps_info) # join everything together and see how much we are missing

test <- aquamaps_depth_cats %>%
  filter(is.na(value)) ## ok so missing 654/23699 = ~3% of depth categories for species 




mode <- function(x) { names(which.max(table(x))) }

results_df <- aquamaps_depth_cats %>%
  mutate(x = TRUE) %>%
  spread(value, x) %>%
  ### Coerce mammals, birds, and reptiles to be pelagic, indicating presence above
  ### the bottom - no air breathing creature is truly benthic!
  mutate(pelagic = ifelse(taxon %in% c('seabirds', 'marine_mammals', 'reptiles'), TRUE, pelagic)) %>%
  mutate(pelagic = 
           case_when(
    Class == "Mammalia" ~ TRUE,
    Class == "Aves" ~ TRUE,
    Class %in% c("Reptilia", "Amphibia") ~ TRUE, 
    TRUE ~ pelagic
  )) %>%
  mutate(depth_position = case_when(benthic & pelagic ~ 'benthopelagic',
                           pelagic ~ 'pelagic',
                           reef ~ 'reef',
                           benthic ~ 'benthic',
                           TRUE ~ 'oops')) %>%
  dplyr::select(species, genus, depth_position, source, DepthPrefMin, DepthPrefMax, FBname, SpeciesID, SpecCode, Class, Family, Order, Phylum, Kingdom) %>%
  distinct() 

genus_mode <- fb_vuln_depth_df %>%
  group_by(genus) %>%
  mutate(nspp = n_distinct(species)) %>%
  group_by(genus, value, nspp) %>%
  summarize(nspp_val = n_distinct(species), .groups = 'drop') %>%
  mutate(pct_spp = nspp_val / nspp) %>%
  filter(pct_spp >= .50) %>%
  filter(nspp > 1) %>%
  dplyr::select(genus, mode_depth_genus = value) %>%
  distinct()


family_mode <- results_df %>%
    group_by(Family) %>%
  summarise(mode_depth_family = mode(depth_position)) %>% # gapfill any missing ones with the mode from their Family
  ungroup()

order_mode <- results_df %>%
    group_by(Order) %>%
  summarise(mode_depth_order = mode(depth_position)) %>% # gapfill any missing ones with the mode from their Family
  ungroup()

class_mode <- results_df %>%
    group_by(Class) %>%
  summarise(mode_depth_class = mode(depth_position)) %>% # gapfill any missing ones with the mode from their Family
  ungroup()

phylum_mode <- results_df %>%
    group_by(Phylum) %>%
  summarise(mode_depth_phylum = mode(depth_position)) %>% # gapfill any missing ones with the mode from their Family
  ungroup()

kingdom_mode <- results_df %>%
    group_by(Kingdom) %>%
  summarise(mode_depth_kingdom = mode(depth_position)) %>% # gapfill any missing ones with the mode from their Family
  ungroup()


  
results_df_gf <- results_df %>% 
  left_join(genus_mode) %>%
    mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_genus, depth_position),
         source = ifelse(depth_position == "oops"|is.na(depth_position), "genus gapfill", source)) %>%
  left_join(family_mode) %>%
  mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_family, depth_position),
         source = ifelse(depth_position == "oops"|is.na(depth_position), "family gapfill", source)) %>%
  left_join(order_mode) %>%
  mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_order, depth_position), 
         source = ifelse(depth_position == "oops"|is.na(depth_position), "order gapfill", source)) %>%
  left_join(class_mode) %>%
  mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_class, depth_position), 
         source = ifelse(depth_position == "oops"|is.na(depth_position), "class gapfill", source))  %>%
    left_join(phylum_mode) %>%
  mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_phylum, depth_position), 
         source = ifelse(depth_position == "oops"|is.na(depth_position), "phylum gapfill", source)) %>%
      left_join(kingdom_mode) %>%
  mutate(depth_position = ifelse(depth_position == "oops"|is.na(depth_position), mode_depth_kingdom, depth_position), 
         source = ifelse(depth_position == "oops"|is.na(depth_position), "kingdom gapfill", source)) %>%
  dplyr::select(species, genus, depth_position, source, DepthPrefMin, DepthPrefMax, FBname, SpeciesID, SpecCode) %>%
  distinct()


dups <- duplicated(results_df_gf$species) # we have some duplicates, as evidenced in the number of rows in our results df. Let's fix those. These are mostly because some species are both benthic and pelagic, and need to be recategorized...

dup_spps <- results_df_gf[dups, ]

dups_results <- results_df_gf %>% 
  filter(species %in% c(dup_spps$species)) %>% 
  pivot_wider(names_from = depth_position, values_from = depth_position) %>%
  mutate(depth_position = case_when(
    benthic == "benthic" & reef == "reef" ~ "reef",
    benthic == "benthic" & pelagic == "pelagic" ~ "benthopelagic",
    TRUE ~ "oops"
  )) %>%
  dplyr::select(-benthic, -reef, -pelagic)

results_df_fin <- results_df_gf %>%
  filter(!(species %in% c(dups_results$species))) %>%
  rbind(dups_results) %>%
  dplyr::select(-FBname) # rebind the fixed dups

length(results_df_fin$species) # 23699 perfect

write.csv(results_df_fin, here("prep/03_prep_spp_habitats/int/spp_depth_info.csv"), row.names = FALSE) # save

test <- read.csv(here("prep/03_prep_spp_habitats/int/spp_depth_info.csv"))


am_spp_depth_df <- aquamaps_spp_df %>%
  left_join(results_df_fin)

length(unique(am_spp_depth_df$SpeciesID)) # 23697... hmmm missing two species? 

setdiff(results_df_fin$species, am_spp_depth_df$species) # [1] "artedidraco mirus"      "careproctus georgianus"

# Both have really small ranges, and no probabilities >0.6, therefore they were cut out. Makes sense. 

write.csv(am_spp_depth_df, file.path(aquamaps_dir, "aquamaps_0.6_depth_prepped.csv"), row.names = FALSE) # save file

```
