==> prep/03_prep_spp_habitats/06a_marine_vuln.Rmd <==
---
title: "Calculate sensitivity values for marine species"
name: "Gage Clawson"
date: "September 25, 2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary 

In this script we assign and gapfill vulnerability values for each species and pressure combination. Each species is vulnerable to either bycatch, biomass removal, or both, depending upon the ingredient which is being fished. These methods are adapted from O'Hara et al., 2024. 

 - Read in vulnerability data from [Butt et al. 2022](https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.3919) and match to Aquamaps species. 
  - ~57% of the Aquamaps species do not have vulnerability scores. We will need to gapfill these. We will gapfill using the same method that the source data used (by genus, family, order, class, etc.)


## References

Butt et al. 2022: A trait-based framework for assessing the vulnerability of marine species to human impacts
  - https://doi.org/10.1002/ecs2.3919 
  
O’Hara, C. et al. (2024) ‘Cumulative human impacts on global marine fauna highlight risk to biological and functional diversity’, PLOS ONE, 19(9), p. e0309788. Available at: https://doi.org/10.1371/journal.pone.0309788.
  
Kaschner, K. et al. (2019) ‘AquaMaps: Predicted range maps for aquatic species’. Available at: https://www.aquamaps.org (Accessed: 11 July 2023).

## Setup 

```{r setup, include=FALSE}

library(tidyverse)
library(tidyr)
library(here)
library(sf)
library(data.table)
library(dtplyr)
library(terra)
library(parallel)
library(strex)
library(janitor)
library(readxl)
library(rfishbase)

source(here("src/directories.R"))

source(here("src/spatial.R"))

aquamaps_dir <- file.path(rdsi_raw_data_dir, "aquamaps")


gall_peters <- "+proj=cea +lon_0=0 +x_0=0 +y_0=0 +lat_ts=45 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
#raster template 
base_rast <- rast(res=0.5)
ext(base_rast) <- c(-180, 180, -90, 90)

```

```{r}
spp_depth_info <- read.csv(here("prep/03_prep_spp_habitats/int/spp_depth_info.csv"))
```


Read in vulnerability data from [Butt et al. 2022](https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.3919) and match to Aquamaps species. 
 - ~57% of the Aquamaps species do not have vulnerability scores. We will need to gapfill these. We will gapfill using the same method that the source data used (by genus and then family.)

```{r}
vuln_taxa <- read.csv(here("prep/03_prep_spp_habitats/data/spp_vuln_butt/vuln_gapfilled_tx.csv"))

unique(vuln_taxa$gapfill) # ok so they only had to gapfill by genus and then family
# [1] "none"   "family" "genus" 

aquamaps_spp <- read.csv(file.path(aquamaps_dir, "GTE10_SPECIESOCCURSUM/speciesoccursum_gte10.csv"))

aquamaps_spp_clean <- aquamaps_spp %>%
  mutate(species = paste0(tolower(Genus), " ", Species)) %>%
  mutate(spp_type = case_when(
    Kingdom == "Plantae" | Phylum %in% c("Ochrophyta") ~ "Marine plant",
    str_detect(SpeciesID, "Fis-") ~ "Fish",
    Class == "Mammalia" ~ "Marine mammal",
    Class == "Aves" ~ "Bird",
    Class %in% c("Reptilia") ~ "Reptiles",
    Class %in% c("Amphibia") ~ "Amphibians",
    Phylum %in% c("Mollusca", "Arthropoda", "Cnidaria", "Echinodermata", "Annelida", "Bryozoa", "Brachiopoda", "Porifera", "Phoronida", "Nemertea", "Chaetognatha", "Platyhelminthes", "Priapulida", "Acanthocephala", "Ctenophora", "Echiura", "Kamptozoa", "Gastrotricha", "Gnathostomulida", "Hemichordata", "Nematoda", "Rotifera", "Sipuncula", "Kamptozoa") | Class %in% c("Appendicularia", "Ascidiacea", "Thaliacea", "Cephalochordata", "Thaliacea") ~ "Invertebrates", 
    Kingdom %in% c("Protozoa", "Bacteria") | Phylum == "Sagenista" ~ "Microorganisms"
  )) %>%
  filter(spp_type != "Microorganisms") # filter these out, we won't assess



test <- aquamaps_spp_clean %>% filter(is.na(taxon)) # 0... good

aquamaps_spp_vuln_taxa <- aquamaps_spp_clean %>%
  left_join(vuln_taxa) %>%
  mutate(class = ifelse(is.na(class), Class, class),
         order = ifelse(is.na(order), Order, order),
         family = ifelse(is.na(family), Family, family),
         genus = ifelse(is.na(genus), Genus, genus)) %>%
  mutate(taxon = case_when(
   is.na(taxon) & Phylum == "Mollusca" ~ "molluscs",
   is.na(taxon) & Class == "Polychaeta" ~ "polychaetes",
   is.na(taxon) & spp_type == "Reptiles" ~ "Reptiles",
   is.na(taxon) & spp_type == "Amphibians" ~ "Amphibians",
    is.na(taxon) & spp_type == "Marine mammal" ~ "Marine mammal",
   is.na(taxon) & spp_type == "Bird" ~ "Bird",
   is.na(taxon) & Phylum == "Porifera" ~ "sponges",
   is.na(taxon) & Class %in% c("Elasmobranchii", "Holocephali") ~ "elasmobranchs",
  is.na(taxon) &  Phylum == "Echinodermata" ~ "echinoderms",
  is.na(taxon) &  Phylum == "Arthropoda" ~ "arthropods",
  is.na(taxon) &  Class == "Cephalopoda" ~ "cephalopods",
   is.na(taxon) & Phylum == "Cnidaria" ~ "cnidaria",
   is.na(taxon) & Class == "Actinopterygii" ~ "fish",
   is.na(taxon) & spp_type == "Marine plant" ~ spp_type,
  TRUE ~ taxon
  )) %>%
  mutate(taxon = case_when(
    is.na(taxon) & spp_type == "Fish" ~ "fish",
    is.na(taxon) & spp_type == "Invertebrates" ~ "molluscs",
    TRUE ~ taxon
  )) %>% 
  mutate(taxon = case_when(
    taxon == "crustacea_arthropods" ~ "arthropods",
    taxon == "marine_mammals" ~ "Marine mammal", 
    taxon == "seabirds" ~ "Bird", 
    taxon == "reptiles" ~ "Reptiles", 
    TRUE ~ taxon
  )) %>% 
  dplyr::select(species, class, order, family, genus, gapfill, vuln_gf_id, SpeciesID, SpecCode, FBname, spp_type, taxon)

unique(aquamaps_spp_vuln_taxa$taxon)
summary(aquamaps_spp_vuln_taxa)

test <- aquamaps_spp_vuln_taxa %>%
  filter(is.na(taxon))

# 13687/23699 = 0.5775349 don't have vulnerability scores... we could gapfill the same way they did by genus and then family. We'll try that and see if it fills everything in. If not, we will gapfill by order and then class as a last resort.  

vuln_scores <- read.csv(here("prep/03_prep_spp_habitats/data/spp_vuln_butt/vuln_gapfilled_score.csv")) %>%
  dplyr::select(vuln_gf_id, bycatch, biomass_removal, habitat_loss_degradation)

aquamaps_vuln_scores <- aquamaps_spp_vuln_taxa %>%
  left_join(vuln_scores) %>%
 mutate(across(class:genus, tolower)) %>%
  group_by(genus) %>%
  mutate(genus_mean_bycatch_gf = mean(bycatch, na.rm = TRUE),
         genus_mean_biomass_gf = mean(biomass_removal, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(family) %>%
  mutate(family_mean_bycatch_gf = mean(bycatch, na.rm = TRUE),
         family_mean_biomass_gf = mean(biomass_removal, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(order) %>%
  mutate(order_mean_bycatch_gf = mean(bycatch, na.rm = TRUE),
         order_mean_biomass_gf = mean(biomass_removal, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(class) %>%
  mutate(class_mean_bycatch_gf = mean(bycatch, na.rm = TRUE),
         class_mean_biomass_gf = mean(biomass_removal, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(spp_type) %>%
  mutate(large_taxa_mean_bycatch_gf = mean(bycatch, na.rm = TRUE),
         large_taxa_mean_biomass_gf = mean(biomass_removal, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(bycatch_gf = 
           case_when(is.na(bycatch) & !is.na(genus_mean_bycatch_gf) ~ genus_mean_bycatch_gf,
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & !is.na(family_mean_bycatch_gf) ~ family_mean_bycatch_gf, 
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & !is.na(order_mean_bycatch_gf) ~ order_mean_bycatch_gf,
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & is.na(order_mean_bycatch_gf) & !is.na(class_mean_bycatch_gf) ~ class_mean_bycatch_gf,
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & is.na(order_mean_bycatch_gf) & is.na(class_mean_bycatch_gf) & !is.na(large_taxa_mean_bycatch_gf) ~ large_taxa_mean_bycatch_gf,
                     spp_type == "Microorganisms" ~ 0,
                     TRUE ~ bycatch
                     )) %>% 
           mutate(biomass_removal_gf = 
           case_when(is.na(biomass_removal) & !is.na(genus_mean_biomass_gf) ~ genus_mean_biomass_gf,
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & !is.na(family_mean_biomass_gf) ~ family_mean_biomass_gf, 
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & !is.na(order_mean_biomass_gf) ~ order_mean_biomass_gf,
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & is.na(order_mean_biomass_gf) & !is.na(class_mean_biomass_gf) ~ class_mean_biomass_gf,
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & is.na(order_mean_biomass_gf) & is.na(class_mean_biomass_gf) & !is.na(large_taxa_mean_biomass_gf) ~ large_taxa_mean_biomass_gf,
                     spp_type == "Microorganisms" ~ 0,
                     TRUE ~ biomass_removal
                     ))  %>%
    mutate(gapfill_flag_bycatch = 
           case_when(is.na(bycatch) & !is.na(genus_mean_bycatch_gf) ~ "genus",
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & !is.na(family_mean_bycatch_gf) ~ "family", 
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & !is.na(order_mean_bycatch_gf) ~ "order",
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & is.na(order_mean_bycatch_gf) & !is.na(class_mean_bycatch_gf) ~ "class",
                     is.na(bycatch) & is.na(genus_mean_bycatch_gf) & is.na(family_mean_bycatch_gf) & is.na(order_mean_bycatch_gf) & is.na(class_mean_bycatch_gf) & !is.na(large_taxa_mean_bycatch_gf) ~ "Larger taxonomic group",
                     spp_type == "Microorganisms" ~ "Microorganism not affected",
                     TRUE ~ "none"
                     )) %>% 
      mutate(gapfill_flag_biomass = 
           case_when(is.na(biomass_removal) & !is.na(genus_mean_biomass_gf) ~ "genus",
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & !is.na(family_mean_biomass_gf) ~ "family", 
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & !is.na(order_mean_biomass_gf) ~ "order",
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & is.na(order_mean_biomass_gf) & !is.na(class_mean_biomass_gf) ~ "class",
                     is.na(biomass_removal) & is.na(genus_mean_biomass_gf) & is.na(family_mean_biomass_gf) & is.na(order_mean_biomass_gf) & is.na(class_mean_biomass_gf) & !is.na(large_taxa_mean_biomass_gf) ~ "Larger taxonomic group",
                     spp_type == "Microorganisms" ~ "Microorganism not affected",
                     TRUE ~ "none"
                     )) %>%
  mutate(bycatch_gf = ifelse(taxon == "corals", habitat_loss_degradation, bycatch_gf),
         biomass_removal_gf = ifelse(taxon == "corals", habitat_loss_degradation, biomass_removal_gf),
         gapfill_flag_biomass = ifelse(taxon == "corals", "none",  gapfill_flag_biomass),
         gapfill_flag_bycatch = ifelse(taxon == "corals", "none", gapfill_flag_bycatch)) %>%
  dplyr::select(species, SpeciesID, SpecCode, FBname, spp_type, taxon, bycatch_gf, biomass_removal_gf) %>%
  mutate(bycatch_gf = ifelse(species == "rhinella marina", 0.75, bycatch_gf),
         biomass_removal_gf = ifelse(species == "rhinella marina", 0.75, biomass_removal_gf)) # assign the missing toad score to be the mean of the reptiles

summary(aquamaps_vuln_scores)
hist(aquamaps_vuln_scores$bycatch_gf) # really concentrated around 0.3/0.4... I bet adding this will reduce affected area by quite a bit.
hist(aquamaps_vuln_scores$biomass_removal_gf) # much higher.. makes sense 
```

Create final info dataset that can be used in the loop in the next script. Here we rescale the vulnerability values to 0, 0.5, and 1, so as to match the terrestrial rescaling.

```{r}
forage_spp <- read.csv(here("data/raw_data/fisheries/forage_fish_list_final.csv")) %>%
    mutate(species = tolower(sci_name)) 
  
trimmings_spp <- read_csv(here("data/raw_data/biomar/trimmings_spp_list.csv")) %>% 
  mutate(species = tolower(scientific_name)) 

allocation <- c("economic", "ge", "mass")
diet <- c("fish-dominant", "plant-dominant")
ingredient <- c("fish meal", "fish oil")
fish_type <- c("forage fish", "trimmings fish")
spp_type <-unique(aquamaps_vuln_scores$spp_type)
taxon_type <- unique(aquamaps_vuln_scores$taxon)


info_df <- expand_grid(
  allocation,
  diet,
  ingredient,
  fish_type,
  SpeciesID = unique(aquamaps_vuln_scores$SpeciesID)
)


overlay_vuln <- info_df %>%
  left_join(aquamaps_vuln_scores) %>%
  left_join(spp_depth_info) %>% 
  mutate(trim_spp = ifelse(species %in% c(trimmings_spp$species), 1, NA),
         for_spp = ifelse(species %in% c(forage_spp$species), 1, NA)) %>%
  mutate(vuln = case_when(
    trim_spp == 1 & for_spp == 1 ~ biomass_removal_gf,
    trim_spp == 1 & is.na(for_spp) & str_detect(fish_type, "forage fish") ~ bycatch_gf,
    trim_spp == 1 & str_detect(fish_type, "trimmings fish") ~ biomass_removal_gf, 
    for_spp == 1 & is.na(trim_spp) & str_detect(fish_type, "trimmings fish") ~ bycatch_gf,
    for_spp == 1 & str_detect(fish_type, "forage fish") ~ biomass_removal_gf,
    TRUE ~ bycatch_gf
  )) %>%
    mutate(catch_type = case_when(
    trim_spp == 1 & for_spp == 1 ~ "catch",
    trim_spp == 1 & is.na(for_spp) & str_detect(fish_type, "forage fish") ~ "bycatch",
    trim_spp == 1 & str_detect(fish_type, "trimmings fish") ~ "catch", 
    for_spp == 1 & is.na(trim_spp) & str_detect(fish_type, "trimmings fish") ~ "bycatch",
    for_spp == 1 & str_detect(fish_type, "forage fish") ~ "catch",
    TRUE ~ "bycatch"
  )) %>%
  mutate(vuln_quartile = case_when(
    vuln < 0.25 ~ 0,
    vuln <= 0.75 & vuln >= 0.25 ~ 0.5, 
    vuln > 0.75 ~ 1
  )) %>%
  dplyr::select(-genus, -source, -DepthPrefMin, -bycatch_gf, -biomass_removal_gf, -FBname, -SpecCode, -vuln) %>%
  mutate(taxon = ifelse(species == "cystophora retroflexa", "Marine plant", taxon))

# write_rds(overlay_vuln, file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds"))


overlay_vuln_new <-overlay_vuln %>%
    mutate(vuln_quartile_plus = case_when(vuln_quartile == 0 ~ 0.1,
                                               vuln_quartile == 0.5 ~ 0.6, 
                                               vuln_quartile == 1 ~ 1,
                                               TRUE ~ vuln_quartile)) %>%
  mutate(vuln_quartile_minus = case_when(vuln_quartile == 0 ~ 0,
                                                vuln_quartile == 0.5 ~ 0.4, 
                                                vuln_quartile == 1 ~ 0.9,
                                               TRUE ~ vuln_quartile))

write_rds(overlay_vuln_new, file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds"))



```
