==> prep/03_prep_spp_habitats/07_marine_overlap_summary.Rmd <==
---
title: "Summarize global marine AOH impacts per species"
name: "Gage Clawson"
date: "September 8, 2023"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Summary 

In this script we prep a dataset of total global impacted habitat per marine species, ingredient, and allocation approach.

## Setup 

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(terra)
library(parallel)
library(strex)
library(janitor)
library(readxl)

source(here("src/directories.R"))

source(here("src/spatial.R"))

aquamaps_dir <- file.path(rdsi_raw_data_dir, "aquamaps")
biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")
```


Read in vulnerability data from [Butt et al. 2022](https://esajournals.onlinelibrary.wiley.com/doi/10.1002/ecs2.3919) and match to Aquamaps species. 
 - ~57% of the Aquamaps species do not have vulnerability scores. We will need to gapfill these. We will gapfill using the same method that the source data used (by genus and then family.)

```{r}
vuln_taxa <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  unite("ingredient", c(fish_type, ingredient), sep = "_") %>%
  mutate(ingredient = case_when(
    ingredient == "trimmings fish_fish meal" ~ "trimmings fish_fish meal, cut offs",
    ingredient == "trimmings fish_fish oil" ~ "trimmings fish_fish oil, cut offs",
    TRUE ~ ingredient
  )) %>%
    pivot_longer(cols = c(vuln_quartile, vuln_quartile_plus, vuln_quartile_minus), names_to = "sensitivity_scenario", values_to = "vulnerability") %>%
  mutate(sensitivity_scenario = case_when(
    sensitivity_scenario == "vuln_quartile" ~ "original",
    sensitivity_scenario == "vuln_quartile_plus" ~ "plus-10",
    sensitivity_scenario == "vuln_quartile_minus" ~ "minus-10",
  )) # include sensitivity scenarios

```

Read in saved dfs from overlap scripts and combine and save 

```{r}

aquamaps_spp <- read.csv(file.path(rdsi_raw_data_dir, "aquamaps/GTE10_SPECIESOCCURSUM/speciesoccursum_gte10.csv"))


impact_df <- purrr::map_df(list.files(file.path(biodiv_dir, "int/aoh_impacts_marine"), full.names = TRUE, recursive = TRUE), readRDS) %>%
  left_join(aquamaps_spp %>% 
              unite(Genus, Species, sep = " ", col = "species") %>%
              mutate(species = tolower(species))
              ) 

test <- impact_df %>% 
  filter(allocation == "economic") %>%
  group_by(diet, sensitivity_scenario) %>%
  summarise(total_impact = sum(impact_total, na.rm = TRUE)) %>%
  ungroup()

length(unique(impact_df$SpeciesID)) # 23029 missing some spp missing.. but they are all polychaetes 

setdiff(unique(aquamaps_spp$SpeciesID), unique(impact_df$SpeciesID))

spp_to_run <- setdiff(unique(aquamaps_spp$SpeciesID), unique(impact_df$SpeciesID)) 

test <- vuln_taxa %>% filter(SpeciesID %in% spp_to_run) 
test2 <- test %>%
  filter(taxon != "polychaetes") %>%
  distinct(species) # ok only missing 2 species with very small range. Likely just no overlap

spp_habs_area <- impact_df %>%
  distinct(species, total_hab_area = hab_area) %>%
  group_by(species) %>%
  summarise(total_hab_area = max(total_hab_area)) # something weird happening with some mollusc spp

write.csv(spp_habs_area, here("prep/03_prep_spp_habitats/int/marine_habitat_areas.csv"), row.names = FALSE)

impact_vuln <- impact_df %>%
  mutate(sensitivity_scenario = ifelse(is.na(sensitivity_scenario), "original", sensitivity_scenario)) %>%
  dplyr::group_by(species, diet, fcr_type, allocation, ingredient, sensitivity_scenario) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>% 
  left_join(vuln_taxa) %>%
  left_join(spp_habs_area) %>%
  dplyr::select(sciname= species, diet, fcr_type, allocation, ingredient, sensitivity_scenario, impact_total, vulnerability, spp_type = taxon, impact_total, total_hab_area)
  
summary(impact_vuln) # no Nas in impact

saveRDS(impact_vuln, here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds"))

```
