==> prep/03_prep_spp_habitats/07_process_mean_sd_summary.Rmd <==
---
title: "Process rasters across taxa or ingredients or both"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(data.table)
library(terra)
library(strex)
library(janitor)
library(sf)
library(tools)
library(glue)
library(tidyterra)
library(rnaturalearth)


source(here("src/directories.R"))

source(here("src/spatial.R"))
source(here("src/fxns.R"))

select <- dplyr::select
setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # set working directory to where this script is located
this_dir <- getwd()

aoh_dir <- file.path(rdsi_raw_data_dir, "AOH_lumbierres/reprojected_csvs/")

base_rast <- rast(res=0.5)
ext(base_rast) <- c(-180, 180, -90, 90)

# read in xy cell_id lookup table for land and ocean
ocean_xy_moll <- readRDS(file.path(this_dir, "data/spatial/moll_template_ocean_xy.rds"))
land_xy_moll <- readRDS(here(this_dir, "data/spatial/moll_template_land_xy.rds"))
all_xy_moll <- rbind(ocean_xy_moll, land_xy_moll)

impact_maps_dir <- file.path(rdsi_dir, "biodiversity_impacts/output")

```

Define functions for processing raster maps of taxon or ingredient aggregated mean impacts 

```{r}
r_to_df <- function(f) {
 # f = "/mnt/rdsi/biodiversity_impacts/output/impact_maps_by_taxon_ingredient/fish-dominant/imp_unwt_forage fish_fish meal_economic_arthropods_sd.tif"
  
  r <- terra::rast(f) 
  crs(r) <- crs(moll_template)
  r <- r %>%
    project(., moll_template, method = "near") # have to project it for some reason first...  
  
  if(str_detect(f, "_sd.tif")){
      df <- data.frame(val = as.vector(values(r)),
                   cell_id = 1:ncell(r)) %>%
    filter(!is.na(val)) 
    
  }else{
  
  df <- data.frame(val = as.vector(values(r)),
                   cell_id = 1:ncell(r)) %>%
    filter(!is.na(val)) %>% 
    filter(val > 0)
}
  
  return(df)
}


combine_taxa_maps <- function(tx_impact_map_df) {

  mean_fs <- tx_impact_map_df %>% filter(p == 'mean') %>% .$f
  sdev_fs <- tx_impact_map_df %>% filter(p == 'sd') %>% .$f
  nspp_fs <- tx_impact_map_df %>% filter(p == 'nspp') %>% .$f

  taxa <- tx_impact_map_df %>% filter(p == "mean") %>% pull(t)
  diet <- tx_impact_map_df %>% filter(p == "mean") %>% pull(d)
  allocation <- tx_impact_map_df %>% filter(p == "mean") %>% pull(a)
  ingredient <- tx_impact_map_df %>% filter(p == "mean") %>% pull(i)

  message(glue('... loading mean maps across taxa and ingredients for {allocation_type} and {diet_type} impacts'))

  mean_df <- parallel::mclapply(mean_fs, mc.cores = 12, FUN = r_to_df) %>%
    setNames(glue("{taxa}_{diet}_{allocation}_{ingredient}")) %>%
    data.table::rbindlist(idcol = 'taxon') %>%
    rename(imp_mean = val) %>%
    dplyr::select(cell_id, taxon, imp_mean)

 message(glue('... loading stdev maps across taxa and ingredients for {allocation_type} and {diet_type} impacts'))

  taxa <- tx_impact_map_df %>% filter(p == "sd") %>% pull(t)
  diet <- tx_impact_map_df %>% filter(p == "sd") %>% pull(d)
  allocation <- tx_impact_map_df %>% filter(p == "sd") %>% pull(a)
  ingredient <- tx_impact_map_df %>% filter(p == "sd") %>% pull(i)
 
 sdev_df <- parallel::mclapply(sdev_fs, mc.cores = 12, FUN = r_to_df) %>%
 setNames(glue("{taxa}_{diet}_{allocation}_{ingredient}")) %>%
    data.table::rbindlist(idcol = 'taxon') %>%
    rename(imp_sd = val) %>%
    dplyr::select(cell_id, taxon, imp_sd)

 message(glue('... loading nspp maps across taxa and ingredients for {allocation_type} and {diet_type} impacts'))

  taxa <- tx_impact_map_df %>% filter(p == "nspp") %>% pull(t)
  diet <- tx_impact_map_df %>% filter(p == "nspp") %>% pull(d)
  allocation <- tx_impact_map_df %>% filter(p == "nspp") %>% pull(a)
  ingredient <- tx_impact_map_df %>% filter(p == "nspp") %>% pull(i)
 
  nspp_df <- parallel::mclapply(nspp_fs, mc.cores = 12, FUN = r_to_df) %>%
 setNames(glue("{taxa}_{diet}_{allocation}_{ingredient}")) %>%
    data.table::rbindlist(idcol = 'taxon') %>%
    rename(imp_nspp = val) %>%
    dplyr::select(cell_id, taxon, imp_nspp)
 
  message('... joining mean, sd, nspp into big-ass dataframe for impacts')
  big_df <- mean_df %>%
    full_join(sdev_df, by = c('taxon', 'cell_id')) %>%
    full_join(nspp_df, by = c('taxon', 'cell_id')) %>%
    filter(!is.na(imp_mean) & !is.na(imp_nspp)) # filter out any NAs here so they don't get caught in the mean calculations
  
  return(big_df)
}

process_mean_rasts <- function(big_df) {
  
  ### Set up for parallel processing
  cell_id_vec <- big_df$cell_id %>% unique()
  n_gps <- 25
  gp_vec <- rep(1:n_gps, length.out = length(cell_id_vec))
  
  ### perform parallel processing
  spp_mean_list <- parallel::mclapply(
    X = 1:n_gps, mc.cores = 12, 
    FUN = function(gp) { ### gp <- 2
      gp_cells <- cell_id_vec[gp_vec == gp]
      message('...processing ', length(gp_cells), ' cells in group ', gp, '...')
      gp_out <- big_df %>%
        filter(cell_id %in% gp_cells) %>%
        group_by(cell_id) %>%
        summarize(imp_mean = (sum(imp_mean * imp_nspp) / sum(imp_nspp))) %>%
        ungroup()
    })
  
  ### gather results
  spp_mean_df <- data.table::rbindlist(spp_mean_list)
  return(spp_mean_df)
}

process_nspp_rasts <- function(big_df){
  
   ### Set up for parallel processing
  cell_id_vec <- big_df$cell_id %>% unique()
  n_gps <- 25
  gp_vec <- rep(1:n_gps, length.out = length(cell_id_vec))
  
  ### perform parallel processing
  spp_nspp_list <- parallel::mclapply(
    X = 1:n_gps, mc.cores = 12, 
    FUN = function(gp) { ### gp <- 2
      gp_cells <- cell_id_vec[gp_vec == gp]
      message('...processing ', length(gp_cells), ' cells in group ', gp, '...')
      gp_out <- big_df %>%
        filter(cell_id %in% gp_cells) %>%
        group_by(cell_id) %>%
        summarize(imp_nspp = sum(imp_nspp, na.rm = TRUE)) %>%
        ungroup()
    })
  
  ### gather results
  spp_nspp_df <- data.table::rbindlist(spp_nspp_list)
  return(spp_nspp_df)
  

}




process_sdev_rasts <- function(big_df) {
  ### Set up for parallel processing
  cell_id_vec <- big_df$cell_id %>% unique()
  n_gps <- 100
  gp_vec <- rep(1:n_gps, length.out = length(cell_id_vec))
  
  ### perform parallel processing
  all_spp_sdev_list <- parallel::mclapply(
    X = 1:n_gps, mc.cores = 12, 
    FUN = function(gp) { ### gp <- round(n_gps / 2)
      gp_cells <- cell_id_vec[gp_vec == gp]
      message('...processing ', length(gp_cells), ' cells in group ', gp, ' of ', n_gps, '...')
      # system.time({
        gp_sdev_out <- big_df %>%
          filter(cell_id %in% gp_cells) %>%
          group_by(cell_id) %>%
          summarize(imp_var = iterated_pooled_var(imp_mean, imp_sd, imp_nspp),
                    imp_sd = sqrt(imp_var)) %>%
          ungroup()
      # })
      return(gp_sdev_out)
    })
  
  ### gather results
  all_spp_sdev <- data.table::rbindlist(all_spp_sdev_list)
  return(all_spp_sdev)
}
```

## Calculate mean and nspp impact maps
  - Across all taxa and ingredients: 1 raster per diet/allocation
  - Across all taxa per ingredient: 1 raster per diet/allocation/ingredient
  - Across all ingredients per taxa: 1 raster per taxon/allocation/diet

```{r assemble taxon impact maps to total mean and nspp maps}
tx_impact_map_df <- data.frame(f = list.files(file.path(impact_maps_dir, "impact_maps_by_taxon_ingredient"), 
                                               pattern = 'imp_unwt_',
                                               full.names = TRUE, recursive = TRUE)) %>%
  mutate(d = str_before_first(str_after_nth(f, "\\/", 6), "\\/"),
         t = str_after_last(str_before_last(file_path_sans_ext(f), "_"), "_"), 
         p = str_after_last(file_path_sans_ext(f), "_"), 
         i = str_before_nth(str_after_nth(file_path_sans_ext(basename(f)), "_", 2), "_", 2), 
         a = str_before_first(str_after_nth(file_path_sans_ext(basename(f)), "_", 4), "_"),)


## calculate means across taxa and ingredients; will end up with 1 raster per diet type and allocation type 

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")

for(allocation_type in allocations){
  for(diet_type in diets){

# allocation_type = "mass"
# diet_type = "plant-dominant"
    
    out_fstem_mean <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_ingredient/{diet_type}/{allocation_type}_mean.tif"))
    out_fstem_nspp <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_ingredient/{diet_type}/{allocation_type}_nspp.tif"))
    
    if(all(file.exists(out_fstem_mean, out_fstem_nspp))){
      message(glue("skipping global mean and nspp calculations for {diet_type}; {allocation_type} because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type)
    
    
  big_df <- combine_taxa_maps(tx_impact_map_df_loop)
    
 process_mean_rasts(big_df) %>%
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_mean) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_mean, overwrite = TRUE)
   
  #plot(log(spp_mean_rast+1))
 
 
  process_nspp_rasts(big_df) %>%       
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_nspp) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_nspp, overwrite = TRUE)

  }
}


## calculate means by taxa across ingredients

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")
taxa <- tx_impact_map_df %>% filter(t != "polychaetes") %>% pull(t) %>% unique

for(allocation_type in allocations){
  for(diet_type in diets){
    for(taxa_type in taxa){

      # allocation_type = "economic"
      # diet_type = "plant-dominant"
      # taxa_type = "polychaetes"
    
    out_fstem_mean <- glue(file.path(impact_maps_dir, "impact_maps_by_taxon_across_ingredient/{diet_type}/{allocation_type}_{taxa_type}_mean.tif"))
    out_fstem_nspp <- glue(file.path(impact_maps_dir, "impact_maps_by_taxon_across_ingredient/{diet_type}/{allocation_type}_{taxa_type}_nspp.tif"))

    
    if(all(file.exists(out_fstem_mean, out_fstem_nspp))){
      message(glue("skipping global mean calculations for {diet_type} diet; {allocation_type} allocation; {taxa_type} taxon because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type, 
             t == taxa_type)
    
    message(glue("processing mean calculations for {diet_type} diet; {allocation_type} allocation; {taxa_type} taxon"))
    
    big_df <- combine_taxa_maps(tx_impact_map_df_loop)
    
 process_mean_rasts(big_df) %>%
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_mean) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_mean, overwrite = TRUE)
 
 
  process_nspp_rasts(big_df) %>%       
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_nspp) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_nspp, overwrite = TRUE)
   
# test <- rast(out_fstem)
  # plot(log(test*100000+1))

   }
  }
}


## calculate means by ingredient across taxa

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")
ingredients <- unique(tx_impact_map_df$i)

for(allocation_type in allocations){
  for(diet_type in diets){
    for(ingredient_type in ingredients){

      # allocation_type = "mass"
      # diet_type = "fish-dominant"
      # ingredient_type = "CropsNES_coconut oil"

    out_fstem_mean <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_by_ingredient/{diet_type}/{allocation_type}_{ingredient_type}_mean.tif"))
    out_fstem_nspp <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_by_ingredient/{diet_type}/{allocation_type}_{ingredient_type}_nspp.tif"))
    
    if(all(file.exists(out_fstem_mean, out_fstem_nspp))){
      message(glue("skipping global mean and nspp calculations for {diet_type} diet; {allocation_type} allocation; {ingredient_type} ingredients because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type, 
             i == ingredient_type)
    
    if(nrow(tx_impact_map_df_loop) == 0){
      
      message("skipping, not a category")
      next()
    }
    
message(glue("processing mean and nspp calculations for {diet_type} diet; {allocation_type} allocation; {ingredient_type} ingredient"))

big_df <- combine_taxa_maps(tx_impact_map_df_loop)

 process_mean_rasts(big_df) %>%       
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_mean) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_mean, overwrite = TRUE)
 
  process_nspp_rasts(big_df) %>%       
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_nspp) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_nspp, overwrite = TRUE)
   
 # test <- rast(out_fstem)
  # plot(log(test*100000+1))

   }
  }
}

```

## Calculate sd and pooled variance impact maps
  - Across all taxa and ingredients: 1 raster per diet/allocation
  - Across all taxa per ingredient: 1 raster per diet/allocation/ingredient
  - Across all ingredients per taxa: 1 raster per taxon/allocation/diet


```{r assemble taxon impact maps to total sdev maps}


tx_impact_map_df <- data.frame(f = list.files(file.path(impact_maps_dir, "impact_maps_by_taxon_ingredient"), 
                                               pattern = 'imp_unwt_',
                                               full.names = TRUE, recursive = TRUE)) %>%
  mutate(d = str_before_first(str_after_nth(f, "\\/", 6), "\\/"),
         t = str_after_last(str_before_last(file_path_sans_ext(f), "_"), "_"), 
         p = str_after_last(file_path_sans_ext(f), "_"), 
         i = str_before_nth(str_after_nth(file_path_sans_ext(basename(f)), "_", 2), "_", 2), 
         a = str_before_first(str_after_nth(file_path_sans_ext(basename(f)), "_", 4), "_"),)


## calculate means across taxa and ingredients; will end up with 1 raster per diet type and allocation type 

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")

for(allocation_type in allocations){
  for(diet_type in diets){

# allocation_type = "mass"
# diet_type = "plant-dominant"
    
    out_fstem_sdev <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_ingredient/{diet_type}/{allocation_type}_sd.tif"))

    if(all(file.exists(out_fstem_sdev))){
      message(glue("skipping global sdev calculations for {diet_type}; {allocation_type} because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type)
    
    
  big_df <- combine_taxa_maps(tx_impact_map_df_loop)
    
  process_sdev_rasts(big_df) %>%
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_sd) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_sdev, overwrite = TRUE)
   
#  rast(out_fstem_sdev) %>% plot()

  }
}


## calculate means by taxa across ingredients

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")
taxa <- tx_impact_map_df %>% filter(t != "polychaetes") %>% pull(t) %>% unique

for(allocation_type in allocations){
  for(diet_type in diets){
    for(taxa_type in taxa){

      # allocation_type = "economic"
      # diet_type = "plant-dominant"
      # taxa_type = "polychaetes"
    
    out_fstem_sdev <- glue(file.path(impact_maps_dir, "impact_maps_by_taxon_across_ingredient/{diet_type}/{allocation_type}_{taxa_type}_sd.tif"))

    
    if(all(file.exists(out_fstem_sdev))){
      message(glue("skipping global sd calculations for {diet_type} diet; {allocation_type} allocation; {taxa_type} taxon because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type, 
             t == taxa_type)
    
    message(glue("processing sd calculations for {diet_type} diet; {allocation_type} allocation; {taxa_type} taxon"))
    
    big_df <- combine_taxa_maps(tx_impact_map_df_loop)
    
 process_sdev_rasts(big_df) %>%
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_sd) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_sdev, overwrite = TRUE)
 
 
# test <- rast(out_fstem_sdev)
  # plot(log(test*100000+1))

   }
  }
}


## calculate means by ingredient across taxa

allocations <- c("mass", "ge", "economic")
diets <- c("fish-dominant", "plant-dominant")
ingredients <- unique(tx_impact_map_df$i)

for(allocation_type in allocations){
  for(diet_type in diets){
    for(ingredient_type in ingredients){

      # allocation_type = "mass"
      # diet_type = "fish-dominant"
      # ingredient_type = "CropsNES_coconut oil"

    out_fstem_sdev <- glue(file.path(impact_maps_dir, "impact_maps_across_taxon_by_ingredient/{diet_type}/{allocation_type}_{ingredient_type}_sd.tif"))
    
    if(all(file.exists(out_fstem_sdev))){
      message(glue("skipping global sd calculations for {diet_type} diet; {allocation_type} allocation; {ingredient_type} ingredients because rasters exist"))
      next()
      
    }

    tx_impact_map_df_loop <- tx_impact_map_df %>%
      filter(a == allocation_type, 
             d == diet_type, 
             i == ingredient_type)
    
    if(nrow(tx_impact_map_df_loop) == 0){
      
      message("skipping, not a category")
      next()
    }
    
message(glue("processing sd calculations for {diet_type} diet; {allocation_type} allocation; {ingredient_type} ingredient"))

big_df <- combine_taxa_maps(tx_impact_map_df_loop)

 process_sdev_rasts(big_df) %>%       
    inner_join(all_xy_moll, by = "cell_id") %>%
      dplyr::select(x, y, imp_sd) %>%
      rast(., type = "xyz", crs = moll_template) %>%
   writeRaster(., out_fstem_sdev, overwrite = TRUE)

   
 # test <- rast(out_fstem_sdev)
  # plot(log(test*100000+1))

   }
  }
}
```


Plots

```{r}
allocation_type <- "economic"

global_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_across_taxon_ingredient"), pattern = allocation_type, recursive = TRUE, full.names = TRUE)

global_maps_mean <- rast(global_maps_list[grep("_mean", global_maps_list)])
names(global_maps_mean) <- str_after_last(str_before_last(global_maps_list[grep("_mean", global_maps_list)], "\\/"), "\\/")

plot(log(global_maps_mean*1000000 + 1))




taxa_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_by_taxon_across_ingredient"), pattern = allocation_type, recursive = TRUE, full.names = TRUE)

taxa_maps_mean_list <- taxa_maps_list[grep("_mean", taxa_maps_list)]

taxa_maps_mean <- rast()  # Create an empty stack to hold the aligned rasters

for (raster_file in taxa_maps_mean_list) {
  # raster_file = taxa_maps_mean_list[[1]]
  
  current_raster <- rast(raster_file)
  aligned_raster <- project(current_raster, moll_template, method = "near")  # Align to the reference raster
  names(aligned_raster) <- str_before_last(str_after_nth(raster_file, "\\/", 6), "_")
  taxa_maps_mean <- c(taxa_maps_mean, aligned_raster)  # Add to the stack
}

taxa_maps_mean
plot(log(taxa_maps_mean*1000000+1))


ingredient_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_across_taxon_by_ingredient"), pattern = allocation_type, recursive = TRUE, full.names = TRUE)

ingredient_maps_mean_list <- ingredient_maps_list[grep("_mean", ingredient_maps_list)]

ingredient_maps_mean <- rast()  # Create an empty stack to hold the aligned rasters

for (raster_file in ingredient_maps_mean_list) {
  # raster_file = taxa_maps_mean_list[[1]]
  
  current_raster <- rast(raster_file)
  aligned_raster <- project(current_raster, moll_template, method = "near")  # Align to the reference raster
  names(aligned_raster) <- str_before_last(str_after_nth(raster_file, "\\/", 6), "_")
  ingredient_maps_mean <- c(ingredient_maps_mean, aligned_raster)  # Add to the stack
}

ingredient_maps_mean
plot(log(ingredient_maps_mean*1000000+1))


```


Do some checking.. view rasters to make sure that they all look ok 

```{r}
allocation_types <- c("mass")
impact_type = "mean"
diet_types <- c("plant-dominant", "fish-dominant")

for(allocation in allocation_types){
  for(diet in diet_types){
  
    # allocation = "ge"
    # diet = "plant-dominant"
    
    files <- list.files(impact_maps_dir, pattern = sprintf("imp_unwt_.*_%s_.*_%s\\.tif", allocation, impact_type), recursive = TRUE, full.names = TRUE)
    

files <- files[grep(diet, files)]
   
     rasters_list <- lapply(files, function(raster_file) {
  raster_obj <- rast(raster_file)
  crs(raster_obj) <- crs(moll_template)
  raster_obj <- project(raster_obj, moll_template)
  return(raster_obj)
})
     
 rasters <- rast(rasters_list)
rast_names <- str_remove_all(str_remove_all(str_after_nth(basename(files), "_", 2), paste0("_", allocation)), paste0("_", impact_type, ".tif")) 
    

names(rasters) <- rast_names
rasters[rasters == 0] <- NA

my_palette <- c("#FED976",  "#FEB24C", "#FD8D3C", "#FC4E2A", "#E31A1C",
"#BD0026", "#800026")
scales::show_col(my_palette)


test <- rasters %>% 
  as.data.frame(xy=TRUE) %>%
  pivot_longer(names_to = "type", values_to = "value", cols = c(3:ncol(.)))
 #  pivot_longer(names_to = "type", values_to = "value", cols = c(3:80))
   # pivot_longer(names_to = "type", values_to = "value", cols = c(3:64))


for(plot_type in unique(test$type)){

  # plot_type = "CropsNES_coconut oil_Bird"
  
ggplot(data = test %>% filter(type == plot_type)) +
  geom_tile(aes(x = x, y = y, fill = value)) + 
    scale_fill_gradientn(colors = my_palette) +
    labs(title = glue("{plot_type}"))

 ggsave(glue(file.path(this_dir, "int/plots/{diet}/{allocation}/{impact_type}_{plot_type}.png")), units=c("in"))

}


  }
}





```
