==> prep/03_prep_spp_habitats/09_combine_and_plot.Rmd <==
---
title: "Summary plots"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Setup

Load libraries

```{r setup, include=FALSE}

library(tidyverse)
library(here)
library(data.table)
library(terra)
library(raster)
library(rnaturalearth)
library(strex)
library(janitor)
library(sf)
library(tools)
library(khroma)
library(scales)
library(fields)
library(patchwork)
library(cowplot)
library(glue)
library(viridis)
library(RColorBrewer)
library(ggpubr)
library(smoothr)

source(here("src/directories.R"))

source(here("src/spatial.R"))

select <- dplyr::select
setwd(dirname(rstudioapi::getSourceEditorContext()$path)) # set working directory to where this script is located
this_dir <- getwd()

biodiv_dir <- file.path(rdsi_dir, "biodiversity_impacts")

impact_maps_dir <- file.path(biodiv_dir, "output")

globe_bbox <- rbind(c(-180, -90), c(-180, 90), 
                      c(180, 90), c(180, -90), c(-180, -90)) 

globe_border <- st_polygon(list(globe_bbox)) %>%
  st_sfc(crs = 4326) %>%
  st_sf(data.frame(rgn = 'globe', geom = .)) %>%
  smoothr::densify(max_distance = 0.5) %>%
  st_transform(crs = crs(moll_template))


```


```{r establish color palettes}

## establish color palette from food footprint project
d_brown <- "#515256"
m_brown <- "#B27B54"
l_brown <- "#BC995F"
green <- "#A8B072"
yellow<- "#EFCE71"
l_yellow <- "#F7F6C1"

jv_pal <- c(l_yellow, yellow, green, l_brown, m_brown, d_brown)

continuous_pal <-  colorRampPalette(jv_pal, space="Lab", bias = 3.5)(10000)
image(volcano, asp=1, col=continuous_pal)

light_gray <- "#F8F9FA"
final_palette <- c(light_gray, continuous_pal)
image(volcano, asp=1, col=final_palette)

palette_red <- c(final_palette, "#B90000")
image(volcano, asp=1, col=palette_red)

# jv_pal_new <- jv_pal[7000:length(continuous_pal)]
# image(volcano, asp=1, col=jv_pal_new)

light_gray <- "#F8F9FA"
final_palette <- c(light_gray, continuous_pal)
scales::show_col(final_palette)

countries_shp <- ne_countries(scale = 110, returnclass = "sf") |> 
  st_transform(crs(moll_template)) # read in countries shapefile


muted <- color("muted")
light <- color("light")
light <- unname(c(light(9)))[c(1,2,3,4,5,6,7,8,9)]
muted <- unname(c(muted(9)))[c(4,5,6,7,8,9)]

muted_palette <- c(muted, light) # this gives the number of raw material... for bar charts at the bottom

# Create a named vector with colors for each spp_type
color_scale <- c(
  "Amphibians" = "#AAAA00", 
  "Bird" = "#EEDD88", 
  "Reptiles" = "#117733", 
  "Terrestrial mammal" = "#DDDDDD", 
  "Marine mammal" = "#44BB99",
  "Marine plant" = "#BBCC33",
  "arthropods" = "#88CCEE",
  "cephalopods" = "#44AA99",
  "cnidaria" = "#EE8866",
  "corals" = "#FFAABB",
  "echinoderms" = "#AA4499",
  "elasmobranchs" = "#99DDFF",
  "fish" = "#77AADD",
  "molluscs" = "#999933",
  "sponges" = "#882255"
)


```



Read in marine and terrestrial dfs and combine

```{r}
marine_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_marine_impact.rds")) %>%
  group_by(sciname, spp_type, diet, fcr_type, allocation, ingredient) %>%
  summarise(vulnerability = mean(vulnerability, na.rm = TRUE), 
            impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total)


terrestrial_overlap <- readRDS(here("prep/03_prep_spp_habitats/output/aoh_terrestrial_mammals_birds_impact.rds")) %>% 
  group_by(sciname, diet, fcr_type, allocation, ingredient, spp_type) %>%
  summarise(vulnerability = 1 - mean(suitability, na.rm = TRUE),
            impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
    dplyr::select(sciname, spp_type, diet, fcr_type, allocation, ingredient, vulnerability, impact_total)

all_overlap <- rbind(marine_overlap, terrestrial_overlap)

spp_impacted <- all_overlap %>%
  group_by(spp_type, diet, allocation, fcr_type) %>%
  summarise(spp_count_impacted = n_distinct(sciname)) %>%
  ungroup()


aquamaps_spp <- readRDS(file.path(rdsi_dir, "biodiversity_impacts/int/spp_vuln_depth_info.rds")) %>%
  distinct(spp_type = taxon, sciname = species)
  
eyres_spp <- readRDS(here("prep/03_prep_spp_habitats/data/iucn/eyres_iucn_spp.rds")) %>%
  dplyr::select(spp_type, sciname = scientific_name) %>%
  mutate(sciname = tolower(sciname)) %>%
  mutate(spp_type = case_when(
    spp_type == "mammals" ~ "Terrestrial mammal",
    spp_type == "amphibians" ~ "Amphibians", 
    spp_type == "reptiles" ~ "Reptiles", 
    spp_type == "birds" ~ "Bird"
  ))

spp_assessed <- rbind(aquamaps_spp, eyres_spp) %>%
  group_by(spp_type) %>%
  summarise(spp_count_assessed = n_distinct(sciname)) %>%
  ungroup()
```

Maps!

Global maps across ingredients and taxon; Delta plots as well

```{r}


## only need to save RDS when impact maps info has changed
allocations <- c("ge", "mass")

for(allocation_type in allocations){

 # allocation_type = "economic"

global_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_across_taxon_ingredient"), pattern = allocation_type, recursive = TRUE, full.names = TRUE)

global_maps_mean <- rast(global_maps_list)
names(global_maps_mean) <- file_path_sans_ext(str_after_nth(global_maps_list, "\\/", 6))


all_df <- as.data.frame(global_maps_mean, xy = TRUE) %>%
  pivot_longer(cols = c(3:ncol(.)), values_to = "value", names_to = "type") %>%
  separate_wider_delim(type, delim = "/", names = c("diet", "fcr_type", "calc_type")) %>%
  mutate(value_log =  ifelse(str_detect(calc_type, "mean"), log(value*1000000 + 1), value)) %>%
  rename(long = x, lat = y)

for(i in unique(all_df$calc_type)){

  # i = "economic_mean"

  loop_df <- all_df %>%
    filter(calc_type == i) 

data.table::fwrite(loop_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{i}.rds.gz")))

}
}



## make plots showing impacts for each scenario 

allocation_types <- c("mass", "economic", "ge")  
calc_types <- c("mean", "nspp", "sd")
types_df <- expand.grid(allocation_types, calc_types) %>%
  unite(new_var, Var1:Var2, sep = "_") %>%
  pull(new_var) %>%
  unique()

diets <- c("fish-dominant", "plant-dominant")
fcrs <- c("regular", "efficienct")

for(type in types_df){
  
  # type = "economic_mean"
  
 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{type}.rds.gz"))

loop_df <-  fread(filepath) 
  
## create maps in separate plots, force common scale between them
maps <- map(.x = diets, 
            .f = function(x, y) 
  ggplot() + 
  geom_tile(data = loop_df  %>% 
              filter(diet == x), aes(x= long, y = lat, fill = value_log)) +  # change to regular value if you don't want log transformed...
      # scale_fill_gradientn(colors = YlOrBr(9), na.value = "white") +  # Set the color for missing values (NA) to white
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
   #  scale_fill_viridis(direction = -1) + 
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
    facet_wrap(~fcr_type) + 
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank())
            )

ggsave(glue(here(this_dir, "output/plots/{type}_plot.png")), plot = plot_grid(plotlist = maps, labels = glue("{diets}:{type}"), label_size = 10, nrow = 2), width = 10, height = 8, bg = "white")


}

test <- loop_df %>% filter(calc_type == "economic_mean")


# create delta (change) plots between each diet scenario

for(type in types_df){
  
  # type = "economic_mean"
 
  # if(file.exists(glue(here(this_dir, "output/plots/delta/{type}_plot.png")))){
  #   cat("exists")
  #   next()
  # }

 filepath <- glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/{type}.rds.gz"))

delta_df <-  fread(filepath) %>%
        dplyr::select(-value_log)

  delta_df <- delta_df %>%
      pivot_wider(names_from = c(diet), values_from = value) %>%
    mutate(`fish-dominant` = ifelse(is.na(`fish-dominant`), 0, `fish-dominant`),
           `plant-dominant` = ifelse(is.na(`plant-dominant`), 0, `plant-dominant`)) %>%
  mutate(delta = (`fish-dominant` - `plant-dominant`)) %>%
    mutate(delta_log = log(delta + 1)) %>%
    mutate(delta_rescale = delta*1000000) %>%
    mutate(delta_cat = case_when(
      delta > 0 ~ "Fish-dominant",
      delta < 0 ~ "Plant-dominant", 
      delta == 0 ~ "No difference"
    ))

  
data.table::fwrite(delta_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/delta/{type}.rds.gz")))


  
if(str_detect(type, "nspp")){
  limit <- c(min(delta_df$delta), max(delta_df$delta))

}else{
limit <- c(quantile(delta_df$delta, 0.05), quantile(delta_df$delta, 0.9))

}

## create maps in separate plots, force common scale between them
p <- ggplot() +
  geom_tile(data = delta_df %>% filter(delta != 0) , aes(x= long, y = lat, fill = delta)) +
    scale_fill_gradient2(low = "blue", mid = "white", high = "red", midpoint = 0, limit = limit, na.value = "white", oob=scales::squish) + # this will make sure anything above the scales limit will be a solid color
  geom_sf(data = countries_shp, fill = NA, colour = "grey") +
    facet_wrap(~fcr_type) +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) +
    labs(caption = str_replace(glue("{type}"), "_", "; "))

p <- ggplot() +
  geom_raster(data = delta_df , aes(x= long, y = lat, fill = delta)) +
  geom_sf(data = countries_shp, fill = NA, colour = "grey") +
        geom_sf(data = globe_border,
            fill = NA, color = 'grey70', 
            size = .1) +
    scale_x_continuous(expand = c(0, 0)) +
    scale_y_continuous(expand = c(0, 0)) +
    theme_void() +
    facet_wrap(~fcr_type) +
    labs(caption = str_replace(glue("{type}"), "_", "; "), 
         fill = 'Δ Extinction Risk') + 
        scale_fill_gradient2(high = '#d01c8b', mid = '#ffffdf', low = '#4dac26', limit = limit, oob=scales::squish) 
      # scale_fill_manual(values = c("#d01c8b", "#ffffdf", "#4dac26")) 
       


  ggsave(plot = p, glue(here(this_dir, "output/plots/delta/{type}_plot_test.png")), width = 6.5, height = 3, dpi = 300)

}

test <- rast("/mnt/rdsi/biodiversity_impacts/output/impact_maps_across_taxon_ingredient/plant-dominant/regular/economic_mean.tif")
test2 <- rast("/mnt/rdsi/biodiversity_impacts/output/impact_maps_across_taxon_ingredient/fish-dominant/regular/economic_mean.tif")

test3 <- test2 - test

test3[test3 == 0] <- NA

plot(test3*1000000)
# seems to match


## zoom in on south america 

delta_df <- data.table::fread((file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_ingredient/delta/economic_mean.rds.gz")))


limit <- c(quantile(delta_df$delta, 0.05), quantile(delta_df$delta, 0.90))

test <- delta_df %>% filter(delta > 0) %>% filter(fcr_type == "regular") %>% dplyr::select(long, lat, delta) %>%
  rast(., type = "xyz", crs = moll_template)
plot(test, col = "red")
  

test_rasterize <- countries_shp %>% filter(continent == "South America") %>%
  rasterize(., moll_template) %>%
  as.data.frame(., xy = TRUE)

## create maps in separate plots, force common scale between them
  ggplot() +
  geom_tile(data = delta_df %>% filter(long %in% c(unique(test_rasterize$x)),
                                       lat %in% c(unique(test_rasterize$y))) , aes(x= long, y = lat, fill = delta)) +
   scale_fill_gradient2(low = "blue", mid = "white", high = "red", midpoint = 0, limit = limit, oob = scales::squish) +
    facet_wrap(~fcr_type) +
                  geom_sf(data = countries_shp %>% filter(continent == "South America"), fill = NA, colour = "grey") +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) +
    labs(caption = "economic_mean")
  
    ggsave(here(this_dir, "output/plots/delta/south_america_compare.png"), width = 10, height = 8, bg = "white")

```

Global maps across ingredients per taxon 
 - there is an if statement within each loop that will skip if the files exists. If you want to rerun all of the files you will need to comment that out. 

```{r}

allocations <- c("economic", "ge", "mass")
calc_types <- c("mean", "sd", "nspp")
taxon <- all_overlap %>% 
  filter(spp_type != "polychaetes") %>% 
  pull(spp_type) %>% unique()

## this part only needs to be run if the source maps changed! 
for(allocation_type in allocations){
    for(taxa_type in taxon){
      for(type in calc_types){

# allocation_type = "economic"
#  taxa_type = "corals"

      out_fstem <- glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{type}.rds.gz"))

      # if(file.exists(out_fstem)){
      # 
      #   message(glue("skipping {allocation_type}_{taxa_type}_{type}"))
      #   next()
      # 
      # }

global_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_by_taxon_across_ingredient"), pattern = glue("{allocation_type}_{taxa_type}"), recursive = TRUE, full.names = TRUE)

stack_rast <- rast()

for(file in global_maps_list){
  # file <- global_maps_list[[1]]

  rast_1 <- rast(file) %>%
    project(., moll_template, method = "near")

  stack_rast <- c(stack_rast, rast_1)
}

names(stack_rast) <- file_path_sans_ext(str_after_nth(global_maps_list, "\\/", 6))


all_df <- as.data.frame(stack_rast, xy = TRUE) %>%
  pivot_longer(cols = c(3:ncol(.)), values_to = "value", names_to = "type") %>%
  filter(!is.na(value)) %>%
  separate_wider_delim(type, delim = "/", names = c("diet", "fcr_type", "taxa_calc_type")) %>%
  separate_wider_delim(taxa_calc_type, delim = "_", names = c("allocation", "taxa", "calc_type")) %>%
  mutate(value_log =  ifelse(str_detect(calc_type, "mean"), log(value*1000000 + 1), value)) %>%
  rename(long = x, lat = y)

for(i in unique(all_df$calc_type)){

  # calc_type = "mean"

  # if(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{i}.rds.gz")) %in% c(unique(done$f))){
  # 
  #   cat("file exists")
  #   next()
  # }

  loop_df <- all_df %>%
    filter(calc_type == i) %>%
    filter(!is.na(value))

  data.table::fwrite(loop_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{i}.rds.gz")))

      }
    }
  }
}
     
  
# make and save plots across ingredients per taxon  

for(allocation_type in allocations){
    for(taxa_type in taxon){
      for(type in calc_types){

        # allocation_type = "economic"
        # taxa_type = "Bird"
        # type = "mean"
        
        all_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{type}.rds.gz"))) %>%
          filter(!is.na(value))
        
diets <- unique(all_df$diet)
  
## create maps in separate plots, force common scale between them
maps <- map(.x = diets, 
            .f = function(x, y) 
  ggplot() + 
  geom_tile(data = all_df  %>% filter(diet == x), aes(x= long, y = lat, fill = value_log)) + 
      # scale_fill_gradientn(colors = YlOrBr(9), na.value = "white") +  # Set the color for missing values (NA) to white
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
   #  scale_fill_viridis(direction = -1) + 
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
    facet_wrap(~fcr_type) +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank())
            )


ggsave(glue(here(this_dir, "output/plots/by_taxa/{allocation_type}_{taxa_type}_{type}_plot.png")), plot = plot_grid(plotlist = maps, labels = glue("{diets}:{allocation_type}:{taxa_type}:{type}"), label_size = 10, nrow = 2), width = 10, height = 8, bg = "white")

    }
  }
}



# save delta plots across ingredients per taxon 

for(allocation_type in allocations){
    for(taxa_type in taxon){
      for(type in calc_types){
  
        # type = "mean"
        # allocation_type = "economic"
        # taxa_type = "arthropods"
  
  # if(file.exists(glue(here(this_dir, "output/plots/delta/by_taxa/{allocation_type}_{taxa_type}_{type}_plot.png")))){
  #   cat("exists")
  #   next()
  # }
  # 
        all_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/{allocation_type}_{taxa_type}_{type}.rds.gz"))) %>%
          filter(!is.na(value))

if(str_detect(type, "mean")){
  
  delta_df <- all_df %>%
      pivot_wider(names_from = c(diet, taxa), values_from = value) %>%
    rename_with(
      ~ case_when(
        str_detect(., "fish-dominant") ~ "fish-dominant",
        str_detect(., "plant-dominant") ~ "plant-dominant",
        TRUE ~ .)
      ) %>%
    mutate(`fish-dominant` = ifelse(is.na(`fish-dominant`), 0, `fish-dominant`),
           `plant-dominant` = ifelse(is.na(`plant-dominant`), 0, `plant-dominant`)) %>%
  mutate(delta = (`fish-dominant` - `plant-dominant`)) %>%
    mutate(delta_log = log(delta + 1)) %>%
    mutate(delta_rescale = delta*1000000)
  
}else{
    delta_df <- all_df %>%
      pivot_wider(names_from = c(diet, taxa), values_from = value) %>%
    rename_with(
      ~ case_when(
        str_detect(., "fish-dominant") ~ "fish-dominant",
        str_detect(., "plant-dominant") ~ "plant-dominant",
        TRUE ~ .)
      ) %>%
          mutate(`fish-dominant` = ifelse(is.na(`fish-dominant`), 0, `fish-dominant`),
           `plant-dominant` = ifelse(is.na(`plant-dominant`), 0, `plant-dominant`)) %>%
  mutate(delta = (`fish-dominant` - `plant-dominant`)) %>%
    mutate(delta_log = log(delta + 1)) %>%
    mutate(delta_rescale = delta*1000000)
}
  
data.table::fwrite(delta_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_across_ingredient/delta/{allocation_type}_{taxa_type}_{type}.rds.gz")))

if(str_detect(type, "nspp")){
  limit <- c(min(delta_df$delta), max(delta_df$delta))

}else{
limit <- c(quantile(delta_df$delta, 0.05), quantile(delta_df$delta, 0.9))
}

## create maps in separate plots, force common scale between them
  ggplot() +
  geom_tile(data = delta_df %>% filter(delta != 0) , aes(x= long, y = lat, fill = delta)) +
    scale_fill_gradient2(low = "blue", mid = "white", high = "red", midpoint = 0,  limit = limit, na.value = "white", oob=scales::squish) + # squish makes sure anything above the scales limit will be a solid red or blue... better to visualize this way since the numbers are so small
              geom_sf(data = countries_shp, fill = NA, colour = "grey") +
    facet_wrap(~fcr_type) +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) +
    labs(caption = glue("{allocation_type}; {taxa_type}; {type}"))


ggsave(glue(here(this_dir, "output/plots/delta/by_taxa/{allocation_type}_{taxa_type}_{type}_plot.png")), width = 10, height = 8, bg = "white")

    }
  }
}
```

Global maps by ingredients across taxon 

```{r}

allocations <- c("economic", "ge", "mass")
calc_types <- c("mean", "sd", "nspp")
ingredients <- all_overlap %>% 
  pull(ingredient) %>% unique()

## commented out code only needs to be run if the source impact maps are updated! 
for(allocation_type in allocations){
  for(ingredient_type in ingredients){

      # ingredient_type = "Soybean_soybean meal"
      # allocation_type = "economic"
      # type = "mean"

global_maps_list <- list.files(file.path(impact_maps_dir, "impact_maps_across_taxon_by_ingredient"), pattern = glue("{allocation_type}_{ingredient_type}"), recursive = TRUE, full.names = TRUE)

stack_rast <- rast()

for(file in global_maps_list){
  # file <- global_maps_list[[1]]

  rast_1 <- rast(file) %>%
    project(., moll_template, method = "near")

  stack_rast <- c(stack_rast, rast_1)
}

names(stack_rast) <- file_path_sans_ext(str_after_nth(global_maps_list, "\\/", 6))


all_df <- as.data.frame(stack_rast, xy = TRUE) %>%
  pivot_longer(cols = c(3:ncol(.)), values_to = "value", names_to = "type") %>%
  separate_wider_delim(type, delim = "/", names = c("diet", "fcr_type", "ingredient_calc_type")) %>%
    separate_wider_delim(ingredient_calc_type, delim = "_", names = c("allocation", "raw_material", "ingredient", "calc_type")) %>%
  mutate(value_log =  ifelse(str_detect(calc_type, "mean"), log(value*1000000 + 1), value)) %>%
  rename(long = x, lat = y) %>%
  filter(!is.na(value)) %>%
  unite("crop_ingredient", c(raw_material:ingredient), sep = "_", remove = FALSE)

for(i in unique(all_df$calc_type)){
  # i = "mean"

  loop_df <- all_df %>%
    filter(calc_type == i) %>%
    filter(!is.na(value))

  data.table::fwrite(loop_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_ingredient/{allocation_type}_{ingredient_type}_{i}.rds.gz")))

   }
  }
}





for(allocation_type in allocations){
    for(ingredient_type in ingredients){
      for(type in calc_types){
        
      # ingredient_type = "Soybean_soybean meal"
      # allocation_type = "economic"
      # type = "mean"
        
                all_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_ingredient/{allocation_type}_{ingredient_type}_{type}.rds.gz"))) 
 
diets <- unique(all_df$diet)
  
## create maps in separate plots, force common scale between them
maps <- map(.x = diets, 
            .f = function(x, y) 
  ggplot() + 
  geom_tile(data = all_df  %>% filter(diet == x), aes(x= long, y = lat, fill = value_log)) + 
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
    facet_wrap(~fcr_type) + 
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank())
            )

ggsave(glue(here(this_dir, "output/plots/by_ingredient/{allocation_type}_{ingredient_type}_{type}_plot.png")), plot = plot_grid(plotlist = maps, labels = glue("{diets}:{allocation_type}:{ingredient_type}:{type}"), label_size = 10, nrow = 2), width = 10, height = 8, bg = "white")

    
    }
  }
}


# make delta plots! 

for(allocation_type in allocations){
    for(ingredient_type in ingredients){
      for(type in calc_types){
  
        # type = "mean"
        # allocation_type = "economic"
        # ingredient_type = "CropsNES_coconut oil"
  
  # if(file.exists(glue(here(this_dir, "output/plots/delta/by_ingredient/{allocation_type}_{ingredient_type}_{type}_plot.png")))){
  #   cat("exists")
  #   next()
  # }
  
        all_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_ingredient/{allocation_type}_{ingredient_type}_{type}.rds.gz"))) %>%
          filter(!is.na(value)) %>%
          dplyr::select(long, lat, fcr_type, allocation, crop_ingredient, calc_type, diet, value) %>%
          filter(crop_ingredient == ingredient_type)

if(str_detect(type, "mean")){
  
  delta_df <- all_df %>%
      pivot_wider(names_from = c(diet, crop_ingredient), values_from = value) %>%
    rename_with(
      ~ case_when(
        str_detect(., "fish-dominant") ~ "fish-dominant",
        str_detect(., "plant-dominant") ~ "plant-dominant",
        TRUE ~ .)
      ) 
  
  if(!("fish-dominant" %in% c(colnames(delta_df)))){
    
    delta_df <- delta_df %>%
      mutate(`fish-dominant` = 0)
  }else if(!("plant-dominant" %in% c(colnames(delta_df)))){
        delta_df <- delta_df %>%
      mutate(`plant-dominant` = 0)
  }
  
  delta_df <- delta_df %>% 
  mutate(`fish-dominant` = coalesce(`fish-dominant`, 0),
         `plant-dominant` = coalesce(`plant-dominant`, 0)) %>%
  mutate(delta = (`fish-dominant` - `plant-dominant`)) %>%
    mutate(delta_log = log(delta + 1)) %>%
    mutate(delta_rescale = delta*1000000)
  
}else{
    delta_df <- all_df %>%
      pivot_wider(names_from = c(diet, crop_ingredient), values_from = value) %>%
    rename_with(
      ~ case_when(
        str_detect(., "fish-dominant") ~ "fish-dominant",
        str_detect(., "plant-dominant") ~ "plant-dominant",
        TRUE ~ .)
      ) 
  
  if(!("fish-dominant" %in% c(colnames(delta_df)))){
    
    delta_df <- delta_df %>%
      mutate(`fish-dominant` = 0)
  }else if(!("plant-dominant" %in% c(colnames(delta_df)))){
        delta_df <- delta_df %>%
      mutate(`plant-dominant` = 0)
  }
  
  delta_df <- delta_df %>% 
  mutate(`fish-dominant` = coalesce(`fish-dominant`, 0),
         `plant-dominant` = coalesce(`plant-dominant`, 0)) %>%
  mutate(delta = (`fish-dominant` - `plant-dominant`)) %>%
    mutate(delta_log = log(delta + 1)) %>%
    mutate(delta_rescale = delta*1000000)
}
  
  
  data.table::fwrite(loop_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_ingredient/{allocation_type}_{ingredient_type}_{type}.rds.gz")))

        
if(str_detect(type, "nspp")){
  limit <- c(min(delta_df$delta), max(delta_df$delta))

}else{
limit <- c(quantile(delta_df$delta, 0.05), quantile(delta_df$delta, 0.9))
}

## create maps in separate plots, force common scale between them
  ggplot() +
  geom_tile(data = delta_df %>% filter(delta != 0) , aes(x= long, y = lat, fill = delta)) +
    scale_fill_gradient2(low = "blue", mid = "white", high = "red", midpoint = 0,  limit = limit, na.value = "white", oob=scales::squish) +
              geom_sf(data = countries_shp, fill = NA, colour = "grey") +
    facet_wrap(~fcr_type) +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) +
    labs(caption = glue("{allocation_type}; {ingredient_type}; {type}"))


ggsave(glue(here(this_dir, "output/plots/delta/by_ingredient/{allocation_type}_{ingredient_type}_{type}_plot.png")), width = 10, height = 8, bg = "white")

    }
  }
}

```

Global maps by raw material across taxon 

```{r}

allocations <- c("economic", "ge", "mass")
calc_types <- c("mean", "sd", "nspp")
ingredients <- all_overlap %>% pull(ingredient) %>% unique()
raw_mats <- all_overlap %>% mutate(raw_material = str_before_first(ingredient, "_")) %>% 
  # filter(!(raw_material %in% c("forage fish", "trimmings fish"))) %>% 
  pull(raw_material) %>% unique()

## commented out code only needs to be run if the source impact maps are updated! 
for(allocation_type in allocations){
  for(raw_mat in raw_mats){
    for(type in calc_types){

    # ingredient_type = "Soybean_soybean meal"
    # allocation_type = "economic"
    # type = "mean"
    #  raw_mat = "CropsNES"

    all_files <- list.files(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_ingredient"), pattern = glue("{allocation_type}_{raw_mat}_.*._{type}"), full.names = TRUE)


all_df <- map_dfr(all_files, ~fread(.x), header = TRUE)

all_df_fin <- all_df %>% 
  mutate(value_log = ifelse(calc_type == "mean", (exp(value) - 1) / 1000000, value)) %>%
  group_by(long, lat, diet, fcr_type, allocation, raw_material, calc_type) %>% 
  summarise(value = sum(value, na.rm = TRUE),
            value_log = sum(value_log, na.rm = TRUE)) %>%
  ungroup()

# test <- all_df_fin %>% group_by(diet, fcr_type, allocation) %>% summarise(sum = sum(value, na.rm = TRUE))


data.table::fwrite(all_df_fin, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_raw_material/{allocation_type}_{raw_mat}_{type}.rds.gz")))


delta_df <- all_df_fin %>%
        pivot_wider(names_from = c(diet), values_from = value) 


  if(!("fish-dominant" %in% c(colnames(delta_df)))){
    
    delta_df <- delta_df %>%
      mutate(`fish-dominant` = 0)
  }else if(!("plant-dominant" %in% c(colnames(delta_df)))){
        delta_df <- delta_df %>%
      mutate(`plant-dominant` = 0)
  }

delta_df <- delta_df %>% 
  mutate(delta = (`fish-dominant` - `plant-dominant`)) 

data.table::fwrite(delta_df, glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_raw_material/delta/{allocation_type}_{raw_mat}_{type}.rds.gz")))

    }
  }
}



for(allocation_type in allocations){
    for(raw_mat in raw_mats){
      for(type in calc_types){
        
      # raw_mat = "forage fish"
      # allocation_type = "economic"
      # type = "nspp"
      
                all_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_raw_material/{allocation_type}_{raw_mat}_{type}.rds.gz"))) 
 
diets <- unique(all_df$diet)
  
## create maps in separate plots, force common scale between them
maps <- map(.x = diets, 
            .f = function(x, y) 
  ggplot() + 
  geom_tile(data = all_df  %>% filter(diet == x), aes(x= long, y = lat, fill = value_log)) + 
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
    facet_wrap(~fcr_type) + 
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank())
            )

ggsave(glue(here(this_dir, "output/plots/by_material/{allocation_type}_{raw_mat}_{type}_plot.png")), plot = plot_grid(plotlist = maps, labels = glue("{diets}:{allocation_type}:{raw_mat}:{type}"), label_size = 10, nrow = 2), width = 10, height = 8, bg = "white")

    
    }
  }
}


# make delta plots! 

for(allocation_type in allocations){
    for(raw_mat in raw_mats){
      for(type in calc_types){
  
        # type = "mean"
        # allocation_type = "economic"
        # raw_mat = "Soybean"
  
  # if(file.exists(glue(here(this_dir, "output/plots/delta/by_ingredient/{allocation_type}_{ingredient_type}_{type}_plot.png")))){
  #   cat("exists")
  #   next()
  # }
  
        delta_df <- fread(glue(file.path(impact_maps_dir, "csvs/impact_maps_across_taxon_by_raw_material/delta/{allocation_type}_{raw_mat}_{type}.rds.gz"))) %>%
          filter(!is.na(delta)) %>%
          dplyr::select(long, lat, fcr_type, allocation, raw_material, calc_type, delta)
  
if(str_detect(type, "nspp")){
  limit <- c(min(delta_df$delta), max(delta_df$delta))
  
}else{
limit <- c(quantile(delta_df$delta, 0.05), quantile(delta_df$delta, 0.9))
}

## create maps in separate plots, force common scale between them
  ggplot() + 
  geom_tile(data = delta_df %>% filter(delta != 0) , aes(x= long, y = lat, fill = delta)) +
    scale_fill_gradient2(low = "blue", mid = "white", high = "red", midpoint = 0,  limit = limit, na.value = "white", oob=scales::squish) +
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
    facet_wrap(~fcr_type) + 
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) + 
    labs(caption = glue("{allocation_type}; {raw_mat}; {type}"))
            

ggsave(glue(here(this_dir, "output/plots/delta/by_material/{allocation_type}_{raw_mat}_{type}_plot.png")), width = 10, height = 8, bg = "white")

    }
  }
}

```



**Sum of impacts**

```{r}
## first lets plot by raw material, a stacked bar chart with total AOH impacted


summary_df <- all_overlap %>% 
  separate_wider_delim(ingredient, delim = "_", names = c("raw_material", "ingredient")) %>%
    group_by(diet, fcr_type, allocation, raw_material, spp_type) %>%
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic")

summary_df$raw_material <- with(summary_df, reorder(raw_material, impact_total, FUN = function(x) sum(x)))

ggplot(summary_df, aes(x = raw_material, y = impact_total, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr_type ~ diet, scales = "free_y") +
  labs(title = "AOH impacted (55236 species)", subtitle = "Economic allocation", x = "Raw material", y = expression(paste("AOH impacted (", km^{2}, ")")), fill = NULL, caption = "A") +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

test <- summary_df %>%
  group_by(diet, fcr_type) %>% 
  summarise(impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup()
```

**Average impacts**

```{r}

## now we want to figure out the same thing, but with average extinction risk? 

# lets just look at mean weighted species impact first 

summary_df <- all_overlap %>%
  filter(allocation == "economic") %>%
    separate_wider_delim(ingredient, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr_type, allocation, raw_material, spp_type) %>%
  summarise(nspp =  n_distinct(sciname), 
            impact_total = sum(impact_total, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(impact_w_mean = impact_total/nspp)
  

summary_df$raw_material <- with(summary_df, reorder(raw_material, impact_w_mean, FUN = function(x) mean(x)))


ggplot(summary_df, aes(x = raw_material, y = impact_w_mean, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr_type ~ diet, scales = "free_y") +
  labs(title = "Number of species weighted mean AOH impacted (55236 species)", subtitle = "Economic allocation", x = "Raw material", y = expression(paste(km^{2}, " impact per species")), fill = NULL, caption = "B") +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed


summary_df$raw_material <- with(summary_df, reorder(raw_material, nspp, FUN = function(x) sum(x)))


ggplot(summary_df, aes(x = raw_material, y = nspp, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr_type ~ diet, scales = "free_y") +
  labs(title = "Number of species impacted", subtitle = "Economic allocation", x = "Raw material", y = "Number of species impacted", fill = NULL) +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

```

**Extract extinction risk metrics from rasters**

```{r}
## lets read in each mean extinction risk raster as data frame and make mean plots per species and raw material for economic allocation

allocations <- c("economic", "mass", "ge")
fcr_types <- c("regular", "efficient")
diet_types <- c("plant-dominant", "fish-dominant")
spp_types <- all_overlap %>% 
  filter(spp_type != "polychaetes") %>% 
 # filter(spp_type == "corals") %>%
  pull(spp_type) %>% unique()

ingredient_types <- all_overlap %>% 
  pull(ingredient) %>% unique()



for(allocation_type in allocations){
  for(fcr_type in fcr_types){
    for(diet_type in diet_types){
      for(spp_type in spp_types){
        for(ingredient_type in ingredient_types){
          
# diet_type = "fish-dominant"
# fcr_type = "regular"
# allocation_type = "economic"
# spp_type = "Amphibians"
# ingredient_type = "Soybean_soybean meal"
          
          if(length(list.files(file.path(glue(impact_maps_dir, "/impact_maps_by_taxon_ingredient/{diet_type}/{fcr_type}")), pattern = glue("{ingredient_type}_{allocation_type}_{spp_type}_mean.tif"), full.names = TRUE)) > 0 ){
        
        
        map_df <- rast(list.files(file.path(glue(impact_maps_dir, "/impact_maps_by_taxon_ingredient/{diet_type}/{fcr_type}")), pattern = glue("{ingredient_type}_{allocation_type}_{spp_type}_mean.tif"), full.names = TRUE)) %>%
          as.data.frame(.) %>%
          group_by() %>%
          summarise(mean_ext_risk = mean(impact_mean, na.rm = TRUE)) %>%
          mutate(allocation = allocation_type, 
                 diet = diet_type, 
                 fcr = fcr_type, 
                 spp_type = spp_type, 
                 ingredient_raw_material = ingredient_type)
        
        
        write.csv(map_df, file.path(glue(impact_maps_dir, "/csvs/impact_maps_by_taxon_ingredient_summary/mean_impacts_{spp_type}_{allocation_type}_{diet_type}_{fcr_type}_{ingredient_type}.csv")), row.names = FALSE)
        
          }else{
            next()
          }
        
        }
      }
    }
  }
}

```

**Mean extinction risk**

```{r}

# now lets join all of those together into one dataframe and save 

csv_files <- list.files(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_ingredient_summary"), pattern = ".csv", full.names = TRUE)

list_of_dfs <- lapply(csv_files, read.csv)

combined_impact_df <- do.call(rbind, list_of_dfs)

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic")

summary_df$raw_material <- with(summary_df, reorder(raw_material, mean_ext_risk, FUN = function(x) mean(x)))


ggplot(summary_df, aes(x = raw_material, y = mean_ext_risk, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Species weighted mean extinction risk (55236 species)", subtitle = "Economic allocation", x = "Raw material", y = "Species weighted mean extinction risk", fill = NULL) +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

## switch the raw materials and species in that plot

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic")

summary_df$spp_type <- with(summary_df, reorder(spp_type, mean_ext_risk, FUN = function(x) sum(x)))

ggplot(summary_df, aes(x = spp_type, y = mean_ext_risk, fill = raw_material)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Species weighted mean extinction risk (55236 species)", caption = "Economic allocation", x = "Raw material", y = "Species weighted mean extinction risk", fill = NULL) +
  # scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

## express average extinction risk as relative to average background rate where a species would go extinct; estimating the normal background rate of extinction voss 2019: https://www.jstor.org/stable/24482652

# median estimates of extinction rates range from 0.023 to 0.135. Typical rates of background extinction may be closer to 0.1. 

# use median rates? 
# plants: 0.053
# arthropods: 0.09
# mammals: 0.023
# molluscs: 0.135
# chordates: 0.064

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(bg_ext_rate = case_when(
    spp_type == "arthropods" ~ 0.09,
    spp_type == "Marine plant" ~ 0.053,
    spp_type %in% c("Marine mammal", "Terrestrial mammal") ~ 0.023,
    spp_type == "molluscs" ~ 0.135, 
    TRUE ~ 0.064
  )) %>%
  filter(allocation == "economic") %>%
  mutate(rel_ext_risk = mean_ext_risk/bg_ext_rate)
  

summary_df$raw_material <- with(summary_df, reorder(raw_material, rel_ext_risk, FUN = function(x) mean(x)))


ggplot(summary_df, aes(x = raw_material, y = rel_ext_risk, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Mean extinction risk relative to background rate of extinction (55236 species)", subtitle = "Economic allocation", x = "Raw material", y = "Extinction risk / Background rate of extinction", fill = NULL) +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

```

**Proportional mean impacts**
Contribution of individual harvest pressures to mean impact across species aoh's by taxon

```{r}

csv_files <- list.files(file.path(impact_maps_dir, "csvs/impact_maps_by_taxon_ingredient_summary"), pattern = ".csv", full.names = TRUE)

list_of_dfs <- lapply(csv_files, read.csv)

combined_impact_df <- do.call(rbind, list_of_dfs)

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr, allocation, raw_material) %>%
  mutate(prop_risk = mean_ext_risk/sum(mean_ext_risk)) %>%
  ungroup()

all_summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  mutate(prop_risk = mean_ext_risk/sum(mean_ext_risk)) %>%
  ungroup() %>% 
  mutate(raw_material = "All") %>%
  rbind(., summary_df) %>%
  filter(allocation == "economic")

custom_order <- c(sort(unique(all_summary_df$raw_material), decreasing = TRUE))

ggplot(all_summary_df, aes(x = reorder(raw_material, match(raw_material, custom_order)), y = prop_risk, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Contribution of each harvest pressure to mean impact across species habitats by taxon", caption = "Economic allocation", x = "Raw material", y = "Proportion of total impact", fill = NULL) +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed


summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  group_by(diet, fcr, allocation, spp_type) %>%
  mutate(prop_risk = mean_ext_risk/sum(mean_ext_risk)) %>%
  ungroup() 


all_summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, fcr, allocation, raw_material) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  mutate(prop_risk = mean_ext_risk/sum(mean_ext_risk)) %>%
  ungroup() %>% 
  mutate(spp_type = "All") %>%
  rbind(., summary_df) %>%
  filter(allocation == "economic")

custom_order <- c(sort(unique(all_summary_df$spp_type), decreasing = TRUE))


ggplot(all_summary_df, aes(x = reorder(spp_type, match(spp_type, custom_order)), y = prop_risk, fill = raw_material)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Contribution of each harvest pressure to mean impact across species habitats by taxon", caption = "Economic allocation", x = "Raw material", y = "Proportion of total impact", fill = NULL) +
  #scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed

```



Plot showing diet percentages 

```{r}
diet_comps <- read.csv(here("prep/02_feed/data/aquaculture_diet_composition.csv")) %>%
  distinct(raw_name, prop_diet, diet) %>%
  filter(prop_diet > 0)

# check sums 
diet_comps %>% filter(diet == "fish-dominant") %>% pull(prop_diet) %>% sum()
diet_comps %>% filter(diet == "plant-dominant") %>% pull(prop_diet) %>% sum()


ingredient_levels <- c( "Fish meal, forage fish",  "Fish oil, forage fish", "Fish meal, trimmings", "Fish oil, trimmings", "Soybean meal", "Corn gluten meal" , "Faba beans" , "Wheat gluten" ,"Wheat" , "Soy protein concentrate",  "Guar meal",  "Pea protein concentrate", "Sunflower meal", "Canola oil", "Linseed oil", "Soy oil", "Pea flour", "Coconut oil", "Micro ingredients")


both_diets <- diet_comps %>% 
  mutate(raw_name = str_replace_all(raw_name, "cut offs", "trimmings")) %>%
      mutate(hjust = if_else(condition = diet == "fish-dominant", true = 1, false = 0)) |> 
    mutate(label_x_pos = if_else(diet =="fish-dominant", true =0.5, false = 2.5)) |> 
  mutate(diet = if_else(condition = diet == "fish-dominant", true = "Fish-\ndominant", false = "Plant-\ndominant")) %>% 
   mutate(diet = factor(diet, levels = rev(c("Plant-\ndominant", "Fish-\ndominant")))) %>%
  filter(raw_name != "other") %>%
  mutate(ingredients = str_to_sentence(raw_name)) %>% 
  mutate(ingredients = factor(ingredients, levels = rev(ingredient_levels))) 


ggplot(data = both_diets,
       aes(x = diet , y = prop_diet, fill = ingredients))+
  geom_col()+
  scale_fill_manual(values =  c("darkgoldenrod4", rev(colorRampPalette(brewer.pal(n=9, name = "Greens"))(14)),  brewer.pal(n=5, name = "Blues")[c(2:5)]))+
  theme_pubr()+
  theme(legend.title = element_blank(),
        legend.text = element_text(size=7),
        text = element_text(size=8),
        legend.position = "right",
        #legend.box.spacing = unit(-0.1, "cm"),
        panel.grid = element_blank(),
        legend.key.size = unit(0.25, "cm")
        )+
  labs(x = "Feed scenario", y = "Proportion")+
  guides(fill = guide_legend(byrow=TRUE, reverse = FALSE, ncol = 1))

ggsave(filename = here("prep/03_prep_spp_habitats/output/plots/feed_composition.jpg"), dpi = 600, width = 10, height = 9, units = "cm")



```

Example disturbance raster 

```{r}
dist_dir <- file.path(biodiv_dir, "int/resampled_ingredient_rasts")

allocations <- c("economic", "mass", "ge")
diets <- c("plant-dominant", "fish-dominant")
fcrs <- c("regular", "efficient")

for(allocation_type in allocations){
  for(diet_type in diets){
    for(fcr_type in fcrs){
      
#       diet_type <- "fish-dominant"
# fcr_type <- "regular"
# allocation_type = "economic"

      stack_dist_fish <- rast(list.files(file.path(dist_dir, glue("{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_.*.A.tif"), full.names = TRUE))

stack_dist_crop <- rast(list.files(file.path(dist_dir, glue("{diet_type}/{fcr_type}")), pattern = glue("{allocation_type}_A.tif"), full.names = TRUE))

stack_dist <- log(c(stack_dist_fish, stack_dist_crop) + 1)

stack_sum_dist <- app(stack_dist, sum, na.rm = TRUE) %>%
  as.data.frame(., xy = TRUE)

 ggplot() + 
  geom_tile(data = stack_sum_dist, aes(x= x, y = y, fill = sum)) + 
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
              geom_sf(data = countries_shp, fill = NA, colour = "grey") + 
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank()) +
   labs(fill = expression(paste(km^2, " logged")), title = glue("Disturbance: {diet_type};{fcr_type};{allocation_type}"))
 
 
  ggsave(filename = here(glue("prep/03_prep_spp_habitats/output/plots/disturbance/{diet_type};{fcr_type};{allocation_type}.png")), width = 10, height = 8, bg = "white")

      
    }
  }
}


```


Let's zoom in on a particular species/region/ingredient so we can show the resolution of our results
 - lets do fish oil, forage fish, economic, mean impacts for finfish taxon under regular fcr scenario and fish-dominant diet

```{r}
europe <- countries_shp %>% filter(continent == "Europe") %>% filter(name != "Russia") %>% filter(name %in% c("Iceland", "United Kingdom", "Norway", "Finland", "Sweden", "Denmark", "Germany", "Netherlands", "Greenland", "Belgium"))
rast <- rast(file.path(biodiv_dir, "output/impact_maps_by_taxon_ingredient/fish-dominant/regular/imp_unwt_forage fish_fish oil_economic_fish_mean.tif")) %>%
  crop(., europe)

# x is from -2000000 to 2000000

plot(rast)

rast_df <- rast %>%
  as.data.frame(., xy = TRUE)

  ggplot() + 
                  geom_sf(data = europe, fill = "gray", colour = "white") + 
  geom_tile(data = rast_df, aes(x= x, y = y, fill = impact_mean)) + 
   scale_fill_gradientn(colors = final_palette, na.value = "white") +
  theme_minimal() +
      theme(
          axis.title = element_blank(),
          axis.ticks = element_blank()) + 
    labs(title = "How does forage fish harvest used for fish oil impact finfish?", fill = "Extinction risk", 
         subtitle = "Economic allocation; Fish-dominant diet; Regular FCRs")
          



```



Look at per country land/eez impacts 

```{r}
allocations <- c("economic", "mass", "ge")
fcr_types <- c("regular", "efficient")
diet_types <- c("plant-dominant", "fish-dominant")
spp_types <- all_overlap %>% 
  filter(spp_type != "polychaetes") %>% 
  pull(spp_type) %>% unique()

ingredient_types <- all_overlap %>% 
  pull(ingredient) %>% unique()

land_eez_rgns <- read.csv(file.path(biodiv_dir, "spatial/rgns_xy_all.csv")) %>% 
  dplyr::select(x, y, rgn_id)

land_eez_rgns_ids <- read.csv(here("data/spatial/output/region_ids.csv"))

for(allocation_type in allocations){
  for(fcr_type in fcr_types){
    for(diet_type in diet_types){
      for(spp_type in spp_types){
        for(ingredient_type in ingredient_types){
          
# diet_type = "fish-dominant"
# fcr_type = "regular"
# allocation_type = "economic"
# spp_type = "Amphibians"
# ingredient_type = "Soybean_soybean meal"
          
          if(length(list.files(file.path(glue(impact_maps_dir, "/impact_maps_by_taxon_ingredient/{diet_type}/{fcr_type}")), pattern = glue("{ingredient_type}_{allocation_type}_{spp_type}_mean.tif"), full.names = TRUE)) > 0 ){
        
        
        map_df <- rast(list.files(file.path(glue(impact_maps_dir, "/impact_maps_by_taxon_ingredient/{diet_type}/{fcr_type}")), pattern = glue("{ingredient_type}_{allocation_type}_{spp_type}_mean.tif"), full.names = TRUE)) %>%
          as.data.frame(., xy = TRUE) %>%
          left_join(land_eez_rgns) %>%
          group_by(rgn_id) %>%
          summarise(mean_ext_risk = mean(impact_mean, na.rm = TRUE)) %>%
          mutate(allocation = allocation_type, 
                 diet = diet_type, 
                 fcr = fcr_type, 
                 spp_type = spp_type, 
                 ingredient_raw_material = ingredient_type) %>%
          ungroup()
        
        
        write.csv(map_df, file.path(glue(impact_maps_dir, "/csvs/by_taxon_ingredient_country_summary/mean_impacts_{spp_type}_{allocation_type}_{diet_type}_{fcr_type}_{ingredient_type}.csv")), row.names = FALSE)
        
          }else{
            next()
          }
        
        }
      }
    }
  }
}


rgn_ids <- read.csv(here("data/spatial/output/region_ids.csv"))

# now lets join all of those together into one dataframe and save 

csv_files <- list.files(file.path(impact_maps_dir, "csvs/by_taxon_ingredient_country_summary"), pattern = ".csv", full.names = TRUE)

list_of_dfs <- lapply(csv_files, read.csv)

combined_impact_df <- do.call(rbind, list_of_dfs) %>%
  left_join(rgn_ids)

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, iso3c, rgn_id, fcr, allocation, spp_type) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic") %>%
  filter(diet == "fish-dominant",
         fcr == "efficient")
  
  top_25_impacts <- summary_df %>% 
   group_by(iso3c) %>%
  summarise(total_impact = sum(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
    arrange(-total_impact) %>%
    mutate(rank = row_number()) %>%
    filter(rank <= 25)
  
  plot_df <- summary_df %>%
    filter(iso3c %in% c(top_25_impacts$iso3c))
  

plot_df$iso3c <- with(plot_df, reorder(iso3c, mean_ext_risk, FUN = function(x) mean(x)))


ggplot(plot_df, aes(x = iso3c, y = mean_ext_risk, fill = spp_type)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(title = "Species weighted mean extinction risk: Top 25", subtitle = "Economic allocation", y = "Species weighted mean extinction risk", fill = NULL) +
  scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed


## do the same thing but make a plot with raw material as the fill value

summary_df <- combined_impact_df %>%
    separate_wider_delim(ingredient_raw_material, delim = "_", names = c("raw_material", "ingredient")) %>%
  group_by(diet, iso3c, rgn_id, fcr, allocation, raw_material) %>%
  summarise(
            mean_ext_risk = mean(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
  filter(allocation == "economic") %>%
  filter(diet == "fish-dominant",
         fcr == "regular")
  
  top_25_impacts <- summary_df %>% 
   group_by(iso3c) %>%
  summarise(total_impact = sum(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup() %>%
    arrange(-total_impact) %>%
    mutate(rank = row_number()) %>%
    filter(rank <= 25)
  
  plot_df <- summary_df %>%
    filter(iso3c %in% c(top_25_impacts$iso3c)) %>%
      group_by(iso3c) %>%
  mutate(total_impact = sum(mean_ext_risk, na.rm = TRUE)) %>%
  ungroup()
  

plot_df$iso3c <- with(plot_df, reorder(iso3c, total_impact, FUN = function(x) mean(x)))

color_scale <- c(
  "Pulses" = "#AAAA00",
  "Wheat" = "#EEDD88",
  "Soybean" = "#117733",
  "Wheat" = "#DDDDDD",
  "trimmings fish" = "#44BB99",
  "forage fish" = "#77AADD",
  "Maize" = "#FFAABB",
  "CropsNES" = "#99DDFF",
  "Rapeseed" = "#882255",
  "Sunflower" = "#AA4499
)


ggplot(plot_df, aes(x = iso3c, y = mean_ext_risk, fill = raw_material)) +
    geom_bar(stat = "identity") +
  facet_grid(fcr ~ diet, scales = "free_y") +
  labs(caption = "Economic allocation", y = "Species weighted mean extinction risk", fill = NULL) +
 scale_fill_manual(values = color_scale) + 
  theme_minimal() + 
  coord_flip()  + 
    theme(axis.text.x = element_text(angle = -90, hjust = 1)) # Adjust the 'angle' value as needed



```
